### 2022-08-05 #552 @santikka
Presubmission inquiry: dynamite - Bayesian Modeling and Causal Inference for Multivariate Longitudinal Data
Submitting Author Name: Santtu Tikka
Submitting Author Github Handle: @santikka
Other Package Authors Github handles: @helske
Repository:  https://github.com/santikka/dynamite
Submission type: Pre-submission

---

```
Package: dynamite
Title: Bayesian Modeling and Causal Inference for Multivariate
    Longitudinal Data
Version: 0.0.1
    person("Santtu", "Tikka", , "santtuth@gmail.com", role = c("aut", "cre"),
    person("Jouni", "Helske", , "jouni.helske@iki.fi", role = "aut",
           comment = c(ORCID = "0000-0001-7130-793X"))
  )
Description: Easy-to-use and efficient interface for
  Bayesian inference of complex panel (time series) data. The package supports
  joint modeling of multiple measurements per individual, time-varying and
  time-invariant effects, and a wide range of discrete and
  continuous distributions. Estimation of the models is carried out via 'Stan'.
    bayesplot,
    checkmate,
    cli,
    data.table (>= 1.14.3),
    glue,
    MASS,
    posterior,
    rstan (>= 2.26.11),
    utils
    covr,
    plm,
    knitr
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd",
    "srr::srr_stats_roclet"))
LazyDataCompression: xz

```

The package performs multivariate bayesian regression modeling for panel data (multiple individuals measured over time). Model estimation is carried out via [Stan](https://mc-stan.org/users/interfaces/rstan).

Yes.

The package is mainly intended for applied researchers working with complex panel data. Panel data is common in many scientific fields, especially in sociology and econometrics. For example, analysing individual-level life-course data is valuable for assessing the effects of policy reforms and other interventions.

Several approaches and R packages exist for panel data analysis, that share some common features with `dynamite`, which we have listed in the package README along with the full feature suite of the package. However, we found that other packages often fall short in crucial aspects, such as only supporting gaussian responses, univariate data or effects constant over time and are thus applicable only in specific constrained scenarios. To the best of our knowledge, there are no other R packages (or software in general) that support all features of `dynamite` simultaneously.

Yes.

Regarding the standards, G1.0 states that:

"G1.0 Statistical Software should list at least one primary reference from published academic literature."

While we certainly have several references in the package vignette, our plan is to also write two scientific papers ourselves: a methodological paper that would serve as the primary reference, and a software paper focusing on the `dynamite` package (possibly in the Journal of Statistical Software). Would it be an issue that this "primary reference" does not yet exist at the review stage?
