### 2022-08-08 #552 @adamhsparks
Hi @santikka,
In reference to your query about the publication requirement. If I understand your inquiry correctly, you have several references for the methods that you are using in the package but this package is the first to combine them, is this correct?
