### 2022-08-08 bcefa08
add print method for dynamiteformula
