+++ man/print.dynamiteformula.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/dynamiteformula.R
+\name{print.dynamiteformula}
+\alias{print.dynamiteformula}
+\title{Print a Dynamite Model Formula}
+\usage{
+\method{print}{dynamiteformula}(x, ...)
+}
+\arguments{
+\item{x}{[\code{dynamiteformula}] The model formula.}
+
+\item{...}{Ignored.}
+}
+\description{
+Print a Dynamite Model Formula
+}
+\examples{
+print(
+  obs(y ~ x, family = "gaussian") +
+  obs(z ~ w, family = "exponential") +
+  aux(d ~ log(y) | init(c(0, 1))) +
+  lags(k = 2) +
+  splines(df = 5) +
+  random(responses = c("y", "z"), correlated = TRUE)
+)
+}
