### 2022-08-08 918df9f
more about aux channels for vignette
