+++ vignettes/dynamite.Rmd
@@ -64,11 +64,11 @@ or thousands, it can be reasonable to assume that the dynamics of the system
 change over time. There are various methods to model time-varying effects in 
 (generalized) linear models based on state space models 
 [@harvey1982; @durbin2012], with various R implementations 
-[@helske2022; @knaus2021; @brodersen2014]. The state space modelling approaches 
-are highly flexible but often computationally demanding due to assumed latent 
-processes for the regression coefficients and an analytically intractable 
-marginal likelihood, so they are most useful in settings with only few 
-subjects. Methods based on the varying coefficients models 
+[@helske2022; @knaus2021; @brodersen2014]. The state space modelling 
+approaches are highly flexible but often computationally demanding due to 
+assumed latent processes for the regression coefficients and an analytically 
+intractable marginal likelihood, so they are most useful in settings with only 
+few subjects. Methods based on the varying coefficients models 
 [@hastie1993; @eubank2004] can be computationally more convenient, especially 
 when the varying coefficients are based on splines. Some of the R 
 implementations of time-varying coefficient models include 
@@ -96,8 +96,8 @@ counterfactual scenarios.
 Consider an individual $i$ with observations 
 $y_{i,t} = (y^1_{i,t}, \ldots, y^c_{i,t},\ldots, y^C_{i,t})$, $t=1,\ldots,T$, 
 $i = 1,\ldots,N$ i.e. at each time point $t$ we have $C$ observations from $N$ 
-individuals. Assume that each element of $y_{i,t}$ can depend on the past values 
-$y_{i,t-1}$, as well as additional exogenous covariates $x_{i,t}$. 
+individuals. Assume that each element of $y_{i,t}$ can depend on the past 
+values $y_{i,t-1}$, as well as additional exogenous covariates $x_{i,t}$. 
 Then, assuming that the elements of $y_{i,t}$ are independent given 
 $y_{i, t-1}$ and $x_{i,t}$, we have
 \[
@@ -184,6 +184,8 @@ distributions for the observations:
 * Gamma: `gamma` (log-link, using mean and shape parameterization).
 * Beta: `beta` (logit-link, using mean and precision parameterization).
 
+## Lagged responses and predictors
+
 The `dynamite` models do not support contemporaneous dependencies in order to 
 avoid complex cyclic dependencies which would make handling missing data, 
 subsequent predictions, and causal inference challenging or impossible. 
@@ -219,6 +221,8 @@ of fixed time points in the model is equal to the largest shift value of any
 observed response variable in the model (defined either via `lag` or the 
 global `lags`).
 
+## Time-varying and time-invariant predictors
+
 The formula within `obs` can also contain an additional special function 
 `varying`, which defines the time-varying part of the model equation, in which 
 case we could write for example 
@@ -244,6 +248,8 @@ time-varying coefficients, same spline basis is used for all coefficients,
 with unique spline coefficients and their standard deviation, 
 as defined in the previous Section.
 
+## Random intercepts
+
 Defining the random intercept term for each group ca be done using 
 component `random` where the first argument defines for which channels the 
 intercept should be added, and second argument defines whether or not these 
@@ -255,10 +261,13 @@ analogously with the typical mixed models. Note however, that currently random
 intercepts are not supported for categorical distribution due to technical 
 reasons.
 
-In addition to declaring response variables via `obs`, we can also use function 
-`aux` to define auxiliary channels which are deterministic functions of other 
-variables. Defining these auxiliary variables explicitly instead of defining 
-them directly on the right-hand side of the formulas (i.e., by using the 
+## Auxiliary channels
+
+In addition to declaring response variables via `obs`, we can also use the 
[…]
