### 2022-08-08 31803c5
tests for missing arguments, add data object name to dynamitefit object for print
