### 2022-08-09 a2d83cb
fix formula.dynamitefit, additional tests
