### 2022-08-09 b04d118
fix examples in docs, fit NSE notes in R CMD check
