+++ man/lags.Rd
@@ -31,10 +31,12 @@ obs(y ~ -1 + varying(~x), family = "gaussian") +
 # A two-channel categorical model with time-invariant predictors
 # here, lag terms are specified manually
 obs(x ~ z + lag(x) + lag(y), family = "categorical") +
+  obs(y ~ z + lag(x) + lag(y), family = "categorical")
 
 # The same categorical model as above, but with the lag terms
 # added using 'lags'
 obs(x ~ z, family = "categorical") +
+  obs(y ~ z, family = "categorical") +
   lags(type = "fixed")
 
 }
