+++ .github/workflows/R-CMD-check.yaml
@@ -48,9 +48,9 @@ jobs:
 
       - uses: r-lib/actions/setup-r@v2
         with:
-          extra-repositories: |
-            https://mc-stan.org/r-packages/
-            https://Rdatatable.gitlab.io/data.table
+          # extra-repositories: |
+          #   https://mc-stan.org/r-packages/
+          #   https://Rdatatable.gitlab.io/data.table
           r-version: ${{ matrix.config.r }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
