+++ DESCRIPTION
@@ -48,5 +48,7 @@ Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd",
 RoxygenNote: 7.2.1
 LazyData: true
 LazyDataCompression: xz
-Additional_repositories: https://mc-stan.org/r-packages/,
-    https://Rdatatable.gitlab.io/data.table
+Remotes:
+  StanHeaders=stan-dev/rstan/StanHeaders@StanHeaders_2.26,
+  rstan=stan-dev/rstan/rstan/rstan@StanHeaders_2.26,
+  Rdatatable/data.table@c4a2085
