+++ .github/workflows/test-coverage.yaml
@@ -20,9 +20,9 @@ jobs:
 
       - uses: r-lib/actions/setup-r@v2
         with:
-          #extra-repositories: |
-          #  https://mc-stan.org/r-packages/
-          #  https://Rdatatable.gitlab.io/data.table
+          extra-repositories: |
+            https://mc-stan.org/r-packages/
+            https://Rdatatable.gitlab.io/data.table
           use-public-rspm: true
 
       - uses: r-lib/actions/setup-r-dependencies@v2
