### 2022-08-22 eb2fb5c
add global_fixed argument to predict
