### 2022-08-23 e5410e9
separate simulated outcomes from fixed predictors during predict
