+++ R/predict.R
@@ -20,6 +20,12 @@
 #'   in terms of memory usage to first remove redundant columns from `newdata`.
 #' @param type \[`character(1)`]\cr Type of prediction,
 #'   `"response"` (default), `"mean"`, or `"link"`.
+#' @param funs \[`list()`]\cr A list containing a summary function for each
+#'   channel of the model that should be applied to the predicted responses
+#'   for each combination of the posterior draws and the time points.
+#'   If empty, the full individual level predictions are retunred. Otherwise,
+#'   this argument will be recycled if fewer functions are provided than
+#'   the number of channels.
 #' @param impute \[`character(1)`]\cr Which imputation scheme to use for
 #'   missing predictor values. Currently supported options are
 #'   no imputation: `"none"` (default), and
@@ -59,6 +65,7 @@
 #' predict(gaussian_example_fit, type = "response", n_draws = 2L)
 predict.dynamitefit <- function(object, newdata = NULL,
                                 type = c("response", "mean", "link"),
+                                funs = list(),
                                 impute = c("none", "locf"),
                                 new_levels = c(
                                   "none", "bootstrap", "gaussian", "original"
@@ -74,6 +81,10 @@ predict.dynamitefit <- function(object, newdata = NULL,
   stopifnot_(
     !inherits(type, "try-error"),
     "Argument {.arg type} must be either \"response\", \"mean\", or \"link\"."
+  )
+  stopifnot_(
+    identical(length(funs), 0L) || all(vapply(funs, is.function, logical(1L))),
+    "Argument {.arg funs} must be a list of functions."
   )
   impute <- onlyif(is.character(impute), tolower(impute))
   impute <- try(match.arg(impute, c("none", "locf")), silent = TRUE)
@@ -98,11 +109,12 @@ predict.dynamitefit <- function(object, newdata = NULL,
     checkmate::test_flag(x = global_fixed),
     "Argument {.arg global_fixed} must be a single {.cls logical} value."
   )
-  predict_dynamitefit(
+  initialize_predict(
     object,
     newdata,
     type,
-    eval_type = "predict",
+    eval_type = "predicted",
+    funs,
     impute,
     new_levels,
     global_fixed,
@@ -113,84 +125,145 @@ predict.dynamitefit <- function(object, newdata = NULL,
 #' Internal Function for Both Predict and Fitted Methods
 #'
 #' @inheritParams predict.dynamitefit
-#' @param eval_type \[`character(1)`]\cr Either `"predict"` or `"fitted"`.
+#' @param eval_type \[`character(1)`]\cr Either `"predicted"` or `"fitted"`.
 #' @noRd
-predict_dynamitefit <- function(object, newdata, type, eval_type,
-                                impute, new_levels, global_fixed, n_draws) {
+initialize_predict <- function(object, newdata, type, eval_type, funs,
+                               impute, new_levels, global_fixed, n_draws) {
   n_draws <- check_ndraws(n_draws, ndraws(object))
   newdata_null <- is.null(newdata)
   newdata <- check_newdata(object, newdata)
   fixed <- as.integer(attr(object$dformulas$all, "max_lag"))
   group_var <- object$group_var
   time_var <- object$time_var
-  dd <- object$dformulas$det
[…]
