+++ R/deterministic.R
@@ -16,13 +16,13 @@ initialize_deterministic <- function(data, dd, dlp, dld, dls) {
     data[, (dlp[[i]]$response) := data[[resp_pred[i]]]]
     data[, (dlp[[i]]$response) := NA]
   }
-  rhs_stoch <- get_predictors(dls)
+  rhs_ls <- get_predictors(dls)
   for (i in seq_along(dls)) {
     stopifnot_(
-      rhs_stoch[i] %in% names(data),
-      "Can't find variable{?s} {.var {rhs_stoch[i]}} in {.arg data}."
+      rhs_ls[i] %in% names(data),
+      "Can't find variable{?s} {.var {rhs_ls[i]}} in {.arg data}."
     )
-    data[, (dls[[i]]$response) := data[[rhs_stoch[i]]]]
+    data[, (dls[[i]]$response) := data[[rhs_ls[i]]]]
     data[, (dls[[i]]$response) := NA]
   }
   for (i in seq_along(dd)) {
@@ -31,23 +31,23 @@ initialize_deterministic <- function(data, dd, dlp, dld, dls) {
     data[, dd[[i]]$response := past]
     data[, dd[[i]]$response := NA]
   }
-  rhs_det <- get_predictors(dld)
-  ro_det <- attr(dld, "rank_order")
+  rhs_ld <- get_predictors(dld)
+  ro_ld <- attr(dld, "rank_order")
   init <- has_past(dld)
-  for (k in ro_det) {
+  for (k in ro_ld) {
     if (init[k]) {
       as_fun <- paste0("as.", dld[[k]]$specials$resp_type)
       past <- do.call(as_fun, args = list(dld[[k]]$specials$past[1L]))
       data[, (dld[[k]]$response) := past]
       data[, (dld[[k]]$response) := NA]
     } else {
-      data[, (dld[[k]]$response) := data[[rhs_det[k]]]]
+      data[, (dld[[k]]$response) := data[[rhs_ld[k]]]]
       data[, (dld[[k]]$response) := NA]
     }
   }
   if (length(dd) > 0L) {
     cl <- get_quoted(dd)
-    res <- try(assign_deterministic(data, cl, 1L), silent = TRUE)
+    res <- try(assign_deterministic(data, 1L, cl), silent = TRUE)
     stopifnot_(
       !inherits(res, "try-error"),
       c(
@@ -65,59 +65,55 @@ initialize_deterministic <- function(data, dd, dlp, dld, dls) {
 #' @param fixed \[`integer(1)`] Number of fixed time points.
 #' @param group_var \[`character(1)`] Grouping variable name.
 #' @noRd
-assign_initial_values <- function(data, dd, dlp, dld, dls,
-                                  idx, fixed, group_var) {
-  n_det <- length(dd)
-  n_lag_pred <- length(dlp)
-  n_lag_det <- length(dld)
-  n_lag_stoch <- length(dls)
-  ro_pred <- ifelse_(
+assign_initial_values <- function(data, idx, dd, dlp, dld, dls,
+                                  fixed, group_var) {
+  ro_lp <- ifelse_(
     is.null(attr(dlp, "rank_order")),
     integer(0L),
     attr(dlp, "rank_order")
   )
-  ro_det <- ifelse_(
+  ro_ld <- ifelse_(
     is.null(attr(dld, "rank_order")),
     integer(0L),
     attr(dld, "rank_order")
   )
-  ro_stoch <- seq_len(n_lag_stoch)
-  resp_pred <- attr(dlp, "original_response")
-  lhs_det <- get_responses(dld)
-  rhs_det <- get_predictors(dld)
-  lhs_stoch <- get_responses(dls)
-  rhs_stoch <- get_predictors(dls)
[…]
