### 2022-08-23 abb7edd
summary version framework ready
