+++ R/predict.R
@@ -23,7 +23,7 @@
 #' @param funs \[`list()`]\cr A list containing a summary function for each
 #'   channel of the model that should be applied to the predicted responses
 #'   for each combination of the posterior draws and the time points.
-#'   If empty, the full individual level predictions are retunred. Otherwise,
+#'   If empty, the full individual level predictions are returned. Otherwise,
 #'   this argument will be recycled if fewer functions are provided than
 #'   the number of channels.
 #' @param impute \[`character(1)`]\cr Which imputation scheme to use for
@@ -218,22 +218,32 @@ initialize_predict <- function(object, newdata, type, eval_type, funs,
       get_responses(object$dformulas$lag_pred)
     )
   )
-  simulated <- newdata[, .SD, .SDcols = resp_draw]
-  observed <- newdata[, .SD, .SDcols = setdiff(names(newdata), resp_draw)]
+  draw_dep <- newdata[, .SD, .SDcols = resp_draw]
+  draw_indep <- newdata[, .SD, .SDcols = setdiff(names(newdata), resp_draw)]
   if (length(funs) > 0L) {
-    predict_summary()
+    predict_summary(
+      object = object,
+      storage = draw_dep,
+      observed = draw_indep,
+      funs = funs,
+      new_levels = new_levels,
+      n_draws = n_draws,
+      fixed = fixed,
+      group_var = group_var,
+      time_var = time_var
+    )
   } else {
     predict_full(
-      object,
-      simulated,
-      observed,
-      type,
-      eval_type,
-      new_levels,
-      n_draws,
-      fixed,
-      group_var,
-      time_var
+      object = object,
+      simulated = draw_dep,
+      observed = draw_indep,
+      type = type,
+      eval_type = eval_type,
+      new_levels = new_levels,
+      n_draws = n_draws,
+      fixed = fixed,
+      group_var = group_var,
+      time_var = time_var
     )
   }
 }
@@ -276,8 +286,8 @@ predict_full <- function(object, simulated, observed,
   time <- observed[[time_var]]
   u_time <- unique(time)
   n_time <- length(u_time)
-  draw_time <- rep(time, each = n_draws)
   time_offset <- which(unique(object$data[[time_var]]) == u_time[1L]) - 1L
+  draw_time <- rep(time, each = n_draws)
   idx <- which(draw_time == u_time[1L]) + (fixed - 1L) * n_draws
   idx_obs <- rep(
     which(time == u_time[1L]) + (fixed - 1L),
@@ -325,6 +335,121 @@ predict_full <- function(object, simulated, observed,
   data.table::setDF(simulated)
   data.table::setDF(observed)
   list(simulated = simulated, observed = observed)
+}
+
+#' Obtain Summarized Predictions
+#'
+#' @inheritParams initialize_predict
+#' @noRd
+predict_summary <- function(object, storage, observed, funs, new_levels,
+                            n_draws, fixed, group_var, time_var) {
+  formulas_stoch <- get_formulas(object$dformulas$stoch)
+  resp_stoch <- get_responses(object$dformulas$stoch)
+  resp_det <- get_responses(object$dformulas$det)
+  lhs_ld <- get_responses(object$dformulas$lag_det)
+  rhs_ld <- get_predictors(object$dformulas$lag_det)
+  lhs_ls <- get_responses(object$dformulas$lag_stoch)
[…]
