### 2022-08-24 71d0a5c
updated documentation, added missing return values
