+++ R/ndraws.R
@@ -1,14 +1,15 @@
 #' Return the Number of Posterior Draws of a `dynamitefit` Object
 #'
-#' @param x \[`dynamitefit`]\cr The model fit object.
 #' @export
 #' @export ndraws
 #' @rdname ndraws-dynamitefit
 #' @aliases ndraws
 #' @method ndraws dynamitefit
+#' @param x \[`dynamitefit`]\cr The model fit object.
 #' @return Number of posterior draws as a single `integer` value.
 #' @examples
 #' ndraws(gaussian_example_fit)
+#'
 ndraws.dynamitefit <- function(x) {
   stopifnot_(
     !missing(x),
