+++ R/predict.R
@@ -21,12 +21,18 @@
 #'   in terms of memory usage to first remove redundant columns from `newdata`.
 #' @param type \[`character(1)`]\cr Type of prediction,
 #'   `"response"` (default), `"mean"`, or `"link"`.
-#' @param funs \[`list()`]\cr A list containing a summary function for each
-#'   channel of the model that should be applied to the predicted responses
-#'   for each combination of the posterior draws and the time points.
-#'   If empty, the full individual level predictions are returned. Otherwise,
-#'   this argument will be recycled if fewer functions are provided than
-#'   the number of channels.
+#' @param funs \[`list()`]\cr A named list whose names should correspond to the
+#'   response variables of the model. Each element of `funs` should be a
+#'   function or a `list` of functions that will be applied to the
+#'   corresponding predicted responses of the channel over the individuals
+#'   for each combination of the posterior draws and time points.
+#'   In other words, the resulting predictions will be averages
+#'   over the individuals. The functions should take the corresponding
+#'   response variable values as their only argument.
+#'   If `funs` is empty, the full individual level predictions are returned
+#'   instead. Note that this argument can only be used
+#'   if there are multiple individuals (i.e., `group` was not `NULL` in the
+#'   `dynamite` call).
 #' @param impute \[`character(1)`]\cr Which imputation scheme to use for
 #'   missing predictor values. Currently supported options are
 #'   no imputation: `"none"` (default), and
@@ -83,10 +89,6 @@ predict.dynamitefit <- function(object, newdata = NULL,
   stopifnot_(
     !inherits(type, "try-error"),
     "Argument {.arg type} must be either \"response\", \"mean\", or \"link\"."
-  )
-  stopifnot_(
-    identical(length(funs), 0L) || all(vapply(funs, is.function, logical(1L))),
-    "Argument {.arg funs} must be a list of functions."
   )
   impute <- onlyif(is.character(impute), tolower(impute))
   impute <- try(match.arg(impute, c("none", "locf")), silent = TRUE)
@@ -220,9 +222,10 @@ initialize_predict <- function(object, newdata, type, eval_type, funs,
       get_responses(object$dformulas$lag_pred)
     )
   )
-  draw_dep <- newdata[, .SD, .SDcols = resp_draw]
+  draw_dep <- newdata[, .SD, .SDcols = c(resp_draw, group_var, time_var)]
   draw_indep <- newdata[, .SD, .SDcols = setdiff(names(newdata), resp_draw)]
   if (length(funs) > 0L) {
+    funs <- parse_funs(object, funs)
     predict_summary(
       object = object,
       storage = draw_dep,
@@ -332,8 +335,9 @@ predict_full <- function(object, simulated, observed,
     assign_deterministic(simulated, idx, cl)
     skip <- FALSE
   }
-  finalize_predict(type, eval_type, resp_stoch, simulated, observed)
+  finalize_predict(type, resp_stoch, simulated, observed)
   simulated[, c(lhs_ld, lhs_ls) := NULL]
+  data.table::setkeyv(simulated, cols = c(".draw", group_var, time_var))
   data.table::setDF(simulated)
   data.table::setDF(observed)
   list(simulated = simulated, observed = observed)
@@ -345,6 +349,16 @@ predict_full <- function(object, simulated, observed,
 #' @noRd
[…]
