### 2022-08-24 5129845
fix predict expansion, test summary variant
