### 2022-08-24 b379088
add expand argument for fitted and predict, fix docs
