+++ tests/testthat/test-predict.R
@@ -5,38 +5,32 @@
 #'
 test_that("predictions are on the same scale as input data", {
   expect_error(
-    out <- predict(gaussian_example_fit, type = "response", n_draws = 1),
+    pred <- predict(gaussian_example_fit, type = "response", n_draws = 1),
     NA
   )
-  pred <- out$simulated
-  obs <- out$observed
-  expect_equal(sd(obs$y), sd(pred$y_new), tolerance = 0.5)
-  expect_equal(mean(obs$y), mean(pred$y_new), tolerance = 0.5)
+  expect_equal(sd(pred$y), sd(pred$y_new), tolerance = 0.5)
+  expect_equal(mean(pred$y), mean(pred$y_new), tolerance = 0.5)
 
   expect_error(
-    out <- fitted(gaussian_example_fit, n_draws = 1),
+    fit <- fitted(gaussian_example_fit, n_draws = 1),
     NA
   )
-  fit <- out$simulated
-  obs <- out$observed
-  expect_equal(sd(obs$y), sd(fit$y_fitted, na.rm = TRUE), tolerance = 0.5)
+  expect_equal(sd(fit$y), sd(fit$y_fitted, na.rm = TRUE), tolerance = 0.5)
   expect_equal(
-    mean(obs$y),
+    mean(fit$y),
     mean(fit$y_fitted, na.rm = TRUE),
     tolerance = 0.5
   )
   expect_error(
-    out <- predict(multichannel_example_fit, type = "response", n_draws = 1),
+    pred <- predict(multichannel_example_fit, type = "response", n_draws = 1),
     NA
   )
-  pred <- out$simulated
-  obs <- out$observed
-  expect_equal(mean(obs$g), mean(pred$g_new), tolerance = 0.5)
-  expect_equal(mean(obs$p), mean(pred$p_new), tolerance = 0.5)
-  expect_equal(mean(obs$b), mean(pred$b_new), tolerance = 0.1)
-  expect_equal(sd(obs$g), sd(pred$g_new), tolerance = 0.5)
-  expect_equal(sd(obs$p), sd(pred$p_new), tolerance = 0.5)
-  expect_equal(sd(obs$b), sd(pred$b_new), tolerance = 0.1)
+  expect_equal(mean(pred$g), mean(pred$g_new), tolerance = 0.5)
+  expect_equal(mean(pred$p), mean(pred$p_new), tolerance = 0.5)
+  expect_equal(mean(pred$b), mean(pred$b_new), tolerance = 0.1)
+  expect_equal(sd(pred$g), sd(pred$g_new), tolerance = 0.5)
+  expect_equal(sd(pred$p), sd(pred$p_new), tolerance = 0.5)
+  expect_equal(sd(pred$b), sd(pred$b_new), tolerance = 0.1)
 })
 
 test_that("prediction works", {
@@ -72,19 +66,17 @@ test_that("prediction works when starting from an arbitrary time point", {
 
   set.seed(1)
   expect_error(
-    out1 <- predict(gaussian_example_fit, newdata = newdata, n_draws = 4),
+    pred1 <- predict(gaussian_example_fit, newdata = newdata, n_draws = 4),
     NA
   )
   set.seed(1)
   expect_error(
-    out2 <- predict(gaussian_example_fit,
+    pred2 <- predict(gaussian_example_fit,
       newdata = newdata |> dplyr::filter(time > 5),
       n_draws = 4
     ),
     NA
   )
-  pred1 <- expand_predict_output(out1)
-  pred2 <- expand_predict_output(out2)
   expect_equal(pred1 |> dplyr::filter(time > 5), pred2)
   fit <- gaussian_example_fit
   fit$data <- fit$data |>
[…]
