+++ R/predict.R
@@ -58,8 +58,17 @@
 #'   If there are no groups, then the options are equivalent.
 #' @param n_draws \[`integer(1)`]\cr Number of posterior samples to use,
 #'   default is `NULL` which uses all samples.
+#' @param expand \[`logical(1)`]\cr If `TRUE` (the default), the output
+#'   is a single `data.frame` containing the original `newdata` and the
+#'   predicted values. Otherwise, a `list` is returned with two components,
+#'   `simulated` and `observed`, where the first contains only the
+#'   predicted values, and the second contains the original `newdata`.
+#'   Setting `expand` to `FALSE` can help conserve memory because `newdata`
+#'   is not replicated `n_draws` times in the output.
+#'   This argument is ignored if `funs` are provided.
 #' @param ... Ignored.
-#' @return A `data.frame` containing the predicted values.
+#' @return A `data.frame` containing the predicted values or a `list` of two
+#'   `data.frames`. See the `expand` argument for details.
 #' @srrstats {G2.3a} Uses match.arg.
 #' @srrstats {RE2.2} Missing responses can be predicted.
 #' @srrstats {G2.13, G2.14, G2.14a, G2.14b, G2.14c, G2.15}
@@ -73,13 +82,12 @@
 #'
 predict.dynamitefit <- function(object, newdata = NULL,
                                 type = c("response", "mean", "link"),
-                                funs = list(),
-                                impute = c("none", "locf"),
+                                funs = list(), impute = c("none", "locf"),
                                 new_levels = c(
                                   "none", "bootstrap", "gaussian", "original"
                                 ),
-                                global_fixed = FALSE,
-                                n_draws = NULL, ...) {
+                                global_fixed = FALSE, n_draws = NULL,
+                                expand = TRUE, ...) {
   stopifnot_(
     !is.null(object$stanfit),
     "No Stan model fit is available."
@@ -113,6 +121,10 @@ predict.dynamitefit <- function(object, newdata = NULL,
     checkmate::test_flag(x = global_fixed),
     "Argument {.arg global_fixed} must be a single {.cls logical} value."
   )
+  stopifnot_(
+    checkmate::test_flag(x = expand),
+    "Argument {.arg expand} must be a single {.cls logical} value."
+  )
   initialize_predict(
     object,
     newdata,
@@ -122,7 +134,8 @@ predict.dynamitefit <- function(object, newdata = NULL,
     impute,
     new_levels,
     global_fixed,
-    n_draws
+    n_draws,
+    expand
   )
 }
 
@@ -131,8 +144,8 @@ predict.dynamitefit <- function(object, newdata = NULL,
 #' @inheritParams predict.dynamitefit
 #' @param eval_type \[`character(1)`]\cr Either `"predicted"` or `"fitted"`.
 #' @noRd
-initialize_predict <- function(object, newdata, type, eval_type, funs,
-                               impute, new_levels, global_fixed, n_draws) {
+initialize_predict <- function(object, newdata, type, eval_type, funs, impute,
+                               new_levels, global_fixed, n_draws, expand) {
   n_draws <- check_ndraws(n_draws, ndraws(object))
   newdata_null <- is.null(newdata)
   newdata <- check_newdata(object, newdata)
@@ -248,18 +261,22 @@ initialize_predict <- function(object, newdata, type, eval_type, funs,
       n_draws = n_draws,
       fixed = fixed,
       group_var = group_var,
-      time_var = time_var
+      time_var = time_var,
+      expand = expand
     )
   }
 }
 
[…]
