### 2022-08-24 6e09f2d
fix funs input, fix globals, add tests for funs and expand
