### 2022-08-25 6df57fc
remove some redundant code
