### 2022-09-16 285631f
fix indexing of shrinkage parameter lambda
