### 2022-09-21 1b37283
allow NULL group in get_ functions, fixes #48
