### 2022-09-21 20dce2c
fix typos and other issues in the vignette
