### 2022-09-21 0d814ac
replace few leftover magritt pipes with native pipe
