### 2022-09-21 e085830
add parameter descriptions for stanblocks_families.R
