### 2022-09-21 5380ca8
added an example of prior predictive simulation
