### 2022-09-21 c38ee08
declarations before statements in Stan code
