+++ R/stanblocks_families.R
@@ -201,32 +201,37 @@ transformed_data_lines_default <- function(y, idt, write_beta, write_delta,
                                            tau_prior_pars = "") {
   i <- rep(seq_len(K_fixed), beta_prior_npars)
   j <- rep(seq_len(beta_prior_npars), each = K_fixed)
-  mtext_fixed <- paste_rows(
-    "matrix[{K_fixed}, {beta_prior_npars}] beta_prior_pars_{y};",
-    "beta_prior_pars_{y}[{i},{j}] = {beta_prior_pars};",
-    .indent = idt(1)
+  declare_beta <- glue::glue(
+    "matrix[{K_fixed}, {beta_prior_npars}] beta_prior_pars_{y};"
+  )
+  state_beta <- glue::glue(
+    "beta_prior_pars_{y}[{i},{j}] = {beta_prior_pars};"
   )
 
   i <- rep(seq_len(K_varying), delta_prior_npars)
   j <- rep(seq_len(delta_prior_npars), each = K_varying)
-  mtext_varying <- paste_rows(
-    "matrix[{K_varying}, {delta_prior_npars}] delta_prior_pars_{y};",
-    "delta_prior_pars_{y}[{i},{j}] = {delta_prior_pars};",
-    .indent = idt(1)
+  declare_delta <- glue::glue(
+    "matrix[{K_varying}, {delta_prior_npars}] delta_prior_pars_{y};"
+  )
+  state_delta <- glue::glue(
+    "delta_prior_pars_{y}[{i},{j}] = {delta_prior_pars};"
   )
 
   i <- rep(seq_len(K_varying), tau_prior_npars)
   j <- rep(seq_len(tau_prior_npars), each = K_varying)
-  mtext_tau <- paste_rows(
-    "matrix[{K_varying}, {tau_prior_npars}] tau_prior_pars_{y};",
-    "tau_prior_pars_{y}[{i},{j}] = {tau_prior_pars};",
-    .indent = idt(1)
+  declare_tau <- glue::glue(
+    "matrix[{K_varying}, {tau_prior_npars}] tau_prior_pars_{y};"
   )
+  state_tau <- glue::glue("tau_prior_pars_{y}[{i},{j}] = {tau_prior_pars};")
 
   paste_rows(
-    onlyif(write_beta, mtext_fixed),
-    onlyif(write_delta, mtext_varying),
-    onlyif(write_tau, mtext_tau),
+    onlyif(write_beta, declare_beta),
+    onlyif(write_delta, declare_delta),
+    onlyif(write_tau, declare_tau),
+    onlyif(write_beta, state_beta),
+    onlyif(write_delta, state_delta),
+    onlyif(write_tau, state_tau),
+    .indent = idt(1),
     .parse = FALSE
   )
 }
@@ -251,44 +256,53 @@ transformed_data_lines_categorical <- function(y, idt, write_alpha, write_beta,
   k <- S - 1L
   i <- rep(seq_len(k), alpha_prior_npars)
   j <- rep(seq_len(alpha_prior_npars), each = k)
-  mtext_alpha <- paste_rows(
-    "matrix[{k}, {alpha_prior_npars}] alpha_prior_pars_{y};",
-    "alpha_prior_pars_{y}[{i},{j}] = {alpha_prior_pars};",
-    .indent = idt(1)
+  declare_alpha <- glue::glue(
+    "matrix[{k}, {alpha_prior_npars}] alpha_prior_pars_{y};"
+  )
+  state_alpha <- glue::glue(
+    "alpha_prior_pars_{y}[{i},{j}] = {alpha_prior_pars};"
   )
 
   k <- K_fixed * (S - 1L)
   i <- rep(seq_len(k), beta_prior_npars)
[…]
