+++ NEWS.md
@@ -4,18 +4,19 @@
     argument, as per issue #48.
   * Fixed some typos and other issues in the vignette raised by @nicholasjclark 
     during the rOpenSci review process.
-  * Added an example of simulating from prior predictive distribution to the 
+  * Added an example on simulating from the prior predictive distribution to the 
     documentation of `predict`.
-  
+  * Declarations now occur before statements in the generated Stan code.
+
 # dynamite 0.0.2
 
-* Improved the memory usage of `predict` and `fitted` by separating the 
-  simulated values from the predictors that are independent of the posterior 
-  draws.
-* Added support for summarized predictions via a new argument `funs`, this
-  can further significantly reduce memory usage when individual level 
-  predictions are not of interest.
+  * Improved the memory usage of `predict` and `fitted` by separating the 
+    simulated values from the predictors that are independent of the posterior 
+    draws.
+  * Added support for summarized predictions via a new argument `funs`, this
+    can further significantly reduce memory usage when individual level 
+    predictions are not of interest.
 
 # dynamite 0.0.1
 
-* First version of `dynamite`
+  * First version of `dynamite`
