+++ DESCRIPTION
@@ -34,6 +34,7 @@ Imports:
     tidyr,
     utils
 Suggests: 
+    cmdstanr,
     covr,
     knitr,
     plm,
