### 2022-09-22 ca46d6f
stanfit NULL if no compiling
