+++ README.Rmd
@@ -152,3 +152,7 @@ For more examples, see the package vignette.
 - [`plm`](https://CRAN.R-project.org/package=plm) contains various methods to estimate linear models for panel data, e.g. the fixed effect models.
 
 - [`lavaan`](https://CRAN.R-project.org/package=lavaan) provides tools for structural equation modelling, and as such can be used to model various panel data models as well.
+
+## Contributing
+
+Contributions are very welcome, see [CONTRIBUTING.md](https://github.com/santikka/dynamite/blob/main/.github/CONTRIBUTING.md) for general guidelines.
