### 2022-09-22 e34b33d
reverse if clause to actually sample something
