### 2022-09-22 ac7ab8e
categorical_logit for old rstan, drop rstan dev-version dependency
