+++ R/stanblocks_families.R
@@ -22,6 +22,7 @@ vectorizable_prior <- function(x) {
 #' Each of the following functions of the form `(stan_block)_lines_(family)`
 #' uses a subset from a  common set of arguments that are documented here.
 #'
+#' @param backend \[`character(1)`]\cr `"rstan"` or `"cmdstanr"`.
 #' @param y \[`character(1)`]\cr The name of the response of the channel.
 #' @param idt \[`function`]\cr An indentation function, see [indenter_()].
 #' @param obs \[`integer()`]\cr A vector of indices indicating non-missing
@@ -224,15 +225,21 @@ transformed_data_lines_default <- function(y, idt, write_beta, write_delta,
   )
   state_tau <- glue::glue("tau_prior_pars_{y}[{i},{j}] = {tau_prior_pars};")
 
-  paste_rows(
-    onlyif(write_beta, declare_beta),
-    onlyif(write_delta, declare_delta),
-    onlyif(write_tau, declare_tau),
-    onlyif(write_beta, state_beta),
-    onlyif(write_delta, state_delta),
-    onlyif(write_tau, state_tau),
-    .indent = idt(1),
-    .parse = FALSE
+  list(
+    declarations = paste_rows(
+      onlyif(write_beta, declare_beta),
+      onlyif(write_delta, declare_delta),
+      onlyif(write_tau, declare_tau),
+      .indent = idt(1),
+      .parse = FALSE
+    ),
+    statements = paste_rows(
+      onlyif(write_beta, state_beta),
+      onlyif(write_delta, state_delta),
+      onlyif(write_tau, state_tau),
+      .indent = idt(1),
+      .parse = FALSE
+    )
   )
 }
 
@@ -292,18 +299,24 @@ transformed_data_lines_categorical <- function(y, idt, write_alpha, write_beta,
     "tau_prior_pars_{y}[{i},{j}] = {tau_prior_pars};"
   )
 
-  paste_rows(
-    mtext,
-    onlyif(write_alpha, declare_alpha),
-    onlyif(write_beta, declare_beta),
-    onlyif(write_delta, declare_delta),
-    onlyif(write_tau, declare_tau),
-    onlyif(write_alpha, state_alpha),
-    onlyif(write_beta, state_beta),
-    onlyif(write_delta, state_delta),
-    onlyif(write_tau, state_tau),
-    .indent = idt(c(0, rep(1, 8))),
-    .parse = FALSE
+  list(
+    declarations = paste_rows(
+      mtext,
+      onlyif(write_alpha, declare_alpha),
+      onlyif(write_beta, declare_beta),
+      onlyif(write_delta, declare_delta),
+      onlyif(write_tau, declare_tau),
+      .indent = idt(c(0, rep(1, 4))),
+      .parse = FALSE
+    ),
+    statements = paste_rows(
+      onlyif(write_alpha, state_alpha),
+      onlyif(write_beta, state_beta),
+      onlyif(write_delta, state_delta),
+      onlyif(write_tau, state_tau),
+      .indent = idt(1),
+      .parse = FALSE
+    )
   )
 }
 
@@ -522,20 +535,18 @@ transformed_parameters_lines_default <- function(y, idt, noncentered, shrinkage,
       .indent = idt(c(1, 2, 2, 2, 2, 1)),
       .parse = FALSE
     )
-    declare_fixed_intercept <- paste_rows(
-      "real alpha_{y};",
-      "vector[{K}] gamma_{y};",
-      .indent = idt(1),
-      .parse = FALSE
-    )
+    declare_fixed_intercept <- "real alpha_{y};"
     state_fixed_intercept <- paste_rows(
+      "{{",
+      "vector[{K}] gamma_{y};",
[…]
