+++ NEWS.md
@@ -9,6 +9,9 @@
   * Declarations now occur before statements in the generated Stan code.
   * Added support for `cmdstanr` via argument `backend` in `dynamite`.
   * Added a link to the contributing guidelines to README.
+  * The package no longer depends on the development version of `rstan`.
+  * `categorical_logit` is now used instead of `categorical_logit_glm` on older
+    `rstan` and `cmdstanr` versions.
 
 # dynamite 0.0.2
 

[…]
