### 2022-09-22 43040a1
fix warning related to rstan/cmdstanr versions, update README
