### 2022-09-23 0b60e08
use rstan::stan_version()
