### 2022-09-23 944cf9a
fix lagged deterministic predictors
