+++ R/dynamite.R
@@ -536,7 +536,7 @@ parse_lags <- function(dformula, data, group_var, time_var, verbose) {
     gl$channels[!gl$stoch],
     sl$channels[!sl$stoch & !sl$pred]
   )
-  attr(dformula_lag_pred, "rank_order") <- order(sl$rank[sl$pred])
+  attr(dformula_lag_pred, "rank_order") <- sl$rank[sl$pred]
   attr(dformula_lag_det, "rank_order") <-
     order(c(gl$rank[!gl$stoch], sl$rank[!sl$stoch & !sl$pred]))
   attr(dformula_lag_pred, "original_response") <- sl$resp[sl$pred]
