### 2022-09-23 c82468d
simplify backend version check
