### 2022-09-23 3b9aa53
fix error message in case multiple observations per time point per id
