### 2022-09-23 8a0d5d2
rename observations in stan, closes #50
