+++ R/stanblocks_families.R
@@ -110,7 +110,7 @@ data_lines_default <- function(y, idt, has_missing, ...) {
 data_lines_categorical <- function(y, idt, has_missing, ...) {
   dtext_def <- data_lines_default(y, idt, has_missing)
   dtext <- paste_rows(
-    "int<lower=0> {y}[T, N];",
+    "int<lower=0> y_{y}[T, N];",
     .indent = idt(1)
   )
   paste_rows(dtext_def, dtext, .parse = FALSE)
@@ -119,7 +119,7 @@ data_lines_categorical <- function(y, idt, has_missing, ...) {
 data_lines_gaussian <- function(y, idt, has_missing, ...) {
   dtext_def <- data_lines_default(y, idt, has_missing)
   dtext <- paste_rows(
-    "matrix[N, T] {y};",
+    "matrix[N, T] y_{y};",
     .indent = idt(1)
   )
   paste_rows(dtext_def, dtext, .parse = FALSE)
@@ -128,7 +128,7 @@ data_lines_gaussian <- function(y, idt, has_missing, ...) {
 data_lines_binomial <- function(y, idt, has_missing, ...) {
   dtext_def <- data_lines_default(y, idt, has_missing)
   dtext <- paste_rows(
-    "int<lower=0> {y}[T, N];",
+    "int<lower=0> y_{y}[T, N];",
     "int<lower=1> trials_{y}[T, N];",
     .indent = idt(1)
   )
@@ -138,7 +138,7 @@ data_lines_binomial <- function(y, idt, has_missing, ...) {
 data_lines_bernoulli <- function(y, idt, has_missing, ...) {
   dtext_def <- data_lines_default(y, idt, has_missing)
   dtext <- paste_rows(
-    "int<lower=0,upper=1> {y}[T, N];",
+    "int<lower=0,upper=1> y_{y}[T, N];",
     .indent = idt(1)
   )
   paste_rows(dtext_def, dtext, .parse = FALSE)
@@ -147,7 +147,7 @@ data_lines_bernoulli <- function(y, idt, has_missing, ...) {
 data_lines_poisson <- function(y, idt, has_missing, has_offset, ...) {
   dtext_def <- data_lines_default(y, idt, has_missing)
   dtext <- paste_rows(
-    "int<lower=0> {y}[T, N];",
+    "int<lower=0> y_{y}[T, N];",
     onlyif(has_offset, "real offset_{y}[T, N];"),
     .indent = idt(1)
   )
@@ -157,7 +157,7 @@ data_lines_poisson <- function(y, idt, has_missing, has_offset, ...) {
 data_lines_negbin <- function(y, idt, has_missing, ...) {
   dtext_def <- data_lines_default(y, idt, has_missing)
   dtext <- paste_rows(
-    "int<lower=0> {y}[T, N];",
+    "int<lower=0> y_{y}[T, N];",
     .indent = idt(1)
   )
   paste_rows(dtext_def, dtext, .parse = FALSE)
@@ -166,7 +166,7 @@ data_lines_negbin <- function(y, idt, has_missing, ...) {
 data_lines_exponential <- function(y, idt, has_missing, ...) {
   dtext_def <- data_lines_default(y, idt, has_missing)
   dtext <- paste_rows(
-    "matrix<lower=0>[N, T] {y};",
+    "matrix<lower=0>[N, T] y_{y};",
     .indent = idt(1)
   )
   paste_rows(dtext_def, dtext, .parse = FALSE)
[…]
