### 2022-09-23 772f298
use same logic for lag det. predictors as actual channels
