### 2022-09-23 cff7707
test debug argument, fix response names in prepare_stan_input, split dynamite function
