### 2022-09-26 e065872
added comments to generated stan code
