+++ R/stanblocks_families.R
@@ -100,9 +100,10 @@ NULL
 
 data_lines_default <- function(y, idt, has_missing, ...) {
   paste_rows(
-    "// Data for response {y}",
+    onlyif(has_missing, "// Missing data indicators"),
     onlyif(has_missing, "int<lower=0> obs_{y}[N, T];"),
     onlyif(has_missing, "int<lower=0> n_obs_{y}[T];"),
+    "// Data",
     .indent = idt(1)
   )
 }
@@ -129,6 +130,7 @@ data_lines_binomial <- function(y, idt, has_missing, ...) {
   dtext_def <- data_lines_default(y, idt, has_missing)
   dtext <- paste_rows(
     "int<lower=0> y_{y}[T, N];",
+    "// Trials for binomial response {y}",
     "int<lower=1> trials_{y}[T, N];",
     .indent = idt(1)
   )
@@ -148,6 +150,7 @@ data_lines_poisson <- function(y, idt, has_missing, has_offset, ...) {
   dtext_def <- data_lines_default(y, idt, has_missing)
   dtext <- paste_rows(
     "int<lower=0> y_{y}[T, N];",
+    "// Offset term",
     onlyif(has_offset, "real offset_{y}[T, N];"),
     .indent = idt(1)
   )
@@ -377,17 +380,22 @@ parameters_lines_default <- function(y, idt, noncentered, lb, has_fixed,
                                      K_fixed, K_varying) {
   oname <- ifelse_(noncentered, "omega_raw_", "omega_")
   paste_rows(
-    onlyif(has_random_intercept, "real<lower=0> sigma_nu_{y};"),
-    onlyif(has_fixed, "vector[{K_fixed}] beta_{y};"),
-    onlyif(has_varying, "matrix[{K_varying}, D] {oname}{y};"),
-    onlyif(has_varying, "vector<lower={lb}>[{K_varying}] tau_{y};"),
+    onlyif(has_random_intercept,
+      "real<lower=0> sigma_nu_{y}; // SD of random intercepts"),
+    onlyif(has_fixed, "vector[{K_fixed}] beta_{y}; // Fixed coefficients"),
+    onlyif(has_varying,
+      "matrix[{K_varying}, D] {oname}{y}; // Spline coefficients"),
+    onlyif(has_varying,
+      "vector<lower={lb}>[{K_varying}] tau_{y}; // SDs for the random walks"),
     ifelse_(
       has_fixed_intercept || has_varying_intercept,
-      "real a_{y};",
+      "real a_{y}; // Mean of the first time point",
       ""
     ),
-    onlyif(has_varying_intercept, "row_vector[D - 1] omega_raw_alpha_{y};"),
-    onlyif(has_varying_intercept, "real<lower={lb}> tau_alpha_{y};"),
+    onlyif(has_varying_intercept,
+      "row_vector[D - 1] omega_raw_alpha_{y}; // Coefficients for alpha"),
+    onlyif(has_varying_intercept,
+      "real<lower={lb}> tau_alpha_{y}; // SD for the random walk"),
     .indent = idt(1)
   )
 }
@@ -398,21 +406,24 @@ parameters_lines_categorical <- function(y, idt, noncentered, lb, has_fixed,
                                          K_fixed, K_varying, S, ...) {
   oname <- ifelse_(noncentered, "omega_raw_", "omega_")
   paste_rows(
-    onlyif(has_fixed, "matrix[{K_fixed}, {S - 1}] beta_{y};"),
-    onlyif(has_varying, "matrix[{K_varying}, D] {oname}{y}[{S - 1}];"),
[…]
