### 2022-09-26 e4fdaff
update NEWS, run-extended
