### 2022-09-27 47c928f
tweak recovery test, change testthat reporter, run-extended
