### 2022-09-27 2dae938
change testthat reporter to handle parallel testing
