### 2022-09-27 1b0a8b7
more verbose status for recovery tests debugging, run-extended
