+++ tests/testthat/test-recovery.R
@@ -24,7 +24,8 @@ test_that("parameters for the linear regression are recovered as with lm", {
   )
   priors$prior <- c("normal(0, 5)", "normal(0, 1)", "exponential(1)")
   fit_dynamite <- dynamite(obs(y ~ x, family = "gaussian"),
-    data = d, time = "time", priors = priors, chains = 1, refresh = 0
+    data = d, time = "time", priors = priors,
+    chains = 1, iter = 2000, refresh = 500
   )
   expect_equal(coef(fit_dynamite)$mean, coef(fit_lm),
     tolerance = 0.01,
@@ -42,7 +43,7 @@ test_that("parameters for the poisson glm are recovered as with glm", {
 
   fit_glm <- glm(y ~ x, data = d, family = poisson)
   fit_dynamite <- dynamite(obs(y ~ x, family = "poisson"),
-    data = d, time = "time", chains = 1, refresh = 0
+    data = d, time = "time", chains = 1, iter = 2000, refresh = 500
   )
   expect_equal(coef(fit_dynamite)$mean, coef(fit_glm),
     tolerance = 0.01,
@@ -61,7 +62,7 @@ test_that("parameters for the binomial glm are recovered as with glm", {
 
   fit_glm <- glm(cbind(y, u - y) ~ x, data = d, family = binomial)
   fit_dynamite <- dynamite(obs(y ~ x + trials(u), family = "binomial"),
-    data = d, time = "time", chains = 1, refresh = 0
+    data = d, time = "time", chains = 1, iter = 2000, refresh = 500
   )
   expect_equal(coef(fit_dynamite)$mean, coef(fit_glm),
     tolerance = 0.01,
@@ -79,7 +80,7 @@ test_that("parameters for the gamma glm are recovered as with glm", {
 
   fit_glm <- glm(y ~ x, data = d, family = Gamma(link = "log"))
   fit_dynamite <- dynamite(obs(y ~ x, family = "gamma"),
-    data = d, time = "time", chains = 1, refresh = 0
+    data = d, time = "time", chains = 1, iter = 2000, refresh = 500
   )
   expect_equal(coef(fit_dynamite)$mean[1:2], coef(fit_glm),
     tolerance = 0.01,
@@ -92,7 +93,7 @@ test_that("parameters for an AR(1) model are recovered as with arima", {
   set.seed(1)
   fit <- dynamite(obs(LakeHuron ~ 1, "gaussian") + lags(),
     data = data.frame(LakeHuron, time = seq_len(length(LakeHuron)), id = 1),
-    "id", "time", chains = 1, refresh = 0
+    "id", "time", chains = 1, iter = 2000, refresh = 500
   )
   fit_arima <- arima(LakeHuron, c(1, 0, 0))
   expect_equal(coef(fit)$mean[2], coef(fit_arima)[1],
