### 2022-09-27 7a649db
try forcing new rstan installation, run-extended
