+++ .github/workflows/R-CMD-check_cranrel.yaml
@@ -1,5 +1,6 @@
 # Workflow derived from https://github.com/r-lib/actions/tree/master/examples
-# Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
+# This is as CMD-check workflow of dynamite
+# but uses CRAN version of rstan and StanHeaders
 on:
   push:
     branches: [main, master]
@@ -20,9 +21,7 @@ jobs:
         config:
           - {os: macOS-latest,   r: 'release'}
           - {os: windows-latest, r: 'release'}
-          - {os: ubuntu-latest,   r: 'devel', http-user-agent: 'release'}
-          - {os: ubuntu-latest,   r: 'release'}
-          - {os: ubuntu-latest,   r: 'oldrel-1'}
+          - {os: ubuntu-latest,  r: 'release'}
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
@@ -49,7 +48,6 @@ jobs:
       - uses: r-lib/actions/setup-r@v2
         with:
           extra-repositories: |
-            https://mc-stan.org/r-packages/
             https://Rdatatable.gitlab.io/data.table
           r-version: ${{ matrix.config.r }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
@@ -64,7 +62,5 @@ jobs:
             RcppParallel
             RcppEigen
             Rcpp
-            rstan
-            StanHeaders
 
       - uses: r-lib/actions/check-r-package@v2
