+++ .github/workflows/R-CMD-check_cranrel.yaml
@@ -7,7 +7,7 @@ on:
   pull_request:
     branches: [main, master]
 
-name: R-CMD-check
+name: R-CMD-check-cranrel
 
 jobs:
   R-CMD-check:
@@ -48,6 +48,7 @@ jobs:
       - uses: r-lib/actions/setup-r@v2
         with:
           extra-repositories: |
+            https://mc-stan.org/r-packages/
             https://Rdatatable.gitlab.io/data.table
           r-version: ${{ matrix.config.r }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
@@ -62,5 +63,7 @@ jobs:
             RcppParallel
             RcppEigen
             Rcpp
+            any::StanHeaders
+            any::rstan
 
       - uses: r-lib/actions/check-r-package@v2
