+++ .github/workflows/R-CMD-check_cranrel.yaml
@@ -56,7 +56,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          cache-version: 3
+          cache-version: 4
           extra-packages: |
             rcmdcheck
             BH
