### 2022-09-27 09e8d9b
don't require cmdstanr in cranrel tests, run-extended
