+++ .github/workflows/R-CMD-check.yaml
@@ -67,4 +67,11 @@ jobs:
             rstan
             StanHeaders
 
+# To be added when we have tests for cmdstanr backend
+# - name: Build CmdStan
+#   run: |
+#     cmdstanr::check_cmdstan_toolchain(fix = TRUE)
+#     cmdstanr::install_cmdstan()
+#   shell: Rscript {0}
+
       - uses: r-lib/actions/check-r-package@v2
