+++ .github/workflows/R-CMD-check_cranrel.yaml
@@ -48,7 +48,6 @@ jobs:
       - uses: r-lib/actions/setup-r@v2
         with:
           extra-repositories: |
-            https://mc-stan.org/r-packages/
             https://Rdatatable.gitlab.io/data.table
           r-version: ${{ matrix.config.r }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
@@ -66,4 +65,8 @@ jobs:
             cran::StanHeaders
             cran::rstan
 
+# Don't check suggested packages as cmdstanr is not available
+
       - uses: r-lib/actions/check-r-package@v2
+        env:
+          _R_CHECK_FORCE_SUGGESTS_: false
