### 2022-09-27 16d14af
don't try to install suggested packages in cranrel workflow
