+++ .github/workflows/R-CMD-check_cranrel.yaml
@@ -56,6 +56,7 @@ jobs:
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           cache-version: 4
+          dependencies: '"hard"'
           extra-packages: |
             rcmdcheck
             BH
