+++ .github/workflows/R-CMD-check_cranrel.yaml
@@ -48,6 +48,7 @@ jobs:
       - uses: r-lib/actions/setup-r@v2
         with:
           extra-repositories: |
+            https://mc-stan.org/r-packages/
             https://Rdatatable.gitlab.io/data.table
           r-version: ${{ matrix.config.r }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
@@ -56,18 +57,13 @@ jobs:
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           cache-version: 4
-          dependencies: '"hard"'
           extra-packages: |
             rcmdcheck
             BH
             RcppParallel
             RcppEigen
             Rcpp
-            cran::StanHeaders
-            cran::rstan
-
-# Don't check suggested packages as cmdstanr is not available
+            cran/StanHeaders
+            cran/rstan
 
       - uses: r-lib/actions/check-r-package@v2
-        env:
-          _R_CHECK_FORCE_SUGGESTS_: false
