### 2022-10-05 da8df9f
return only the stan code and not model object
