### 2022-10-05 7c70463
fix .draw column in as.data.frame
