+++ NEWS.md
@@ -14,6 +14,9 @@
     `rstan` and `cmdstanr` versions.
   * Random intercepts with `random` now also support centered parameterization.
   * Added more comments to the generated Stan code.
+  * `get_code` applied to fitted model now correctly returns only the model 
+    code and not the stanmodel object.
+  * Fixed the `.draw` column of the `as_data_frame` output.
 
 # dynamite 0.0.2
 
