### 2022-10-06 1e70adb
added LOO method
