+++ R/predict_helpers.R
@@ -45,7 +45,8 @@ check_newdata <- function(object, newdata) {
 #' @param newdata \[`data.frame`]\cr The data to be used for prediction.
 #' @param data \[`data.frame`]\cr The original data used to fit the model.
 #' @param type \[`character`]\cr Either `"response"`, `"mean"`, `"link"`.
-#' @param eval_type \[`character(1)`]\cr Either `"predict"` or `"fitted"`
+#' @param eval_type \[`character(1)`]\cr Either `"predict"`, `"fitted"`, or
+#'   `loglik.`
 #' @param families_stoch \[`list`]\cr of `dynamitefamily` object.
 #' @param resp_stoch \[`character`]\cr A vector of response variables
 #' @param categories \[`list`]\cr Response variable categories
@@ -135,17 +136,21 @@ parse_newdata <- function(dformula, newdata, data, type, eval_type,
     newdata[, (clear_names) := NULL]
   }
   drop_unused(dformula, newdata, group_var, time_var)
-  type <- ifelse_(identical(eval_type, "fitted"), "fitted", type)
+  type <- ifelse_(eval_type %in% c("fitted", "loglik"), eval_type, type)
   # create separate column for each level of categorical response variables
   for (i in seq_along(resp_stoch)) {
     resp <- resp_stoch[i]
-    if (type %in% c("mean", "link", "fitted")) {
-      pred_col <- ifelse_(
-        is_categorical(families_stoch[[i]]),
-        glue::glue("{resp}_{type}_{categories[[resp]]}"),
-        glue::glue("{resp}_{type}")
-      )
-      newdata[, (pred_col) := NA_real_]
+    if (identical(type, "loglik")) {
+      newdata[, (glue::glue("{resp}_loglik")) := NA_real_]
+    } else {
+      if (type %in% c("mean", "link", "fitted")) {
+        pred_col <- ifelse_(
+          is_categorical(families_stoch[[i]]),
+          glue::glue("{resp}_{type}_{categories[[resp]]}"),
+          glue::glue("{resp}_{type}")
+        )
+        newdata[, (pred_col) := NA_real_]
+      }
     }
     newdata[, (glue::glue("{resp}_store")) := newdata[[resp]]]
   }
@@ -533,7 +538,8 @@ prepare_eval_envs <- function(object, simulated, observed,
 #' @param resp \[`character(1)`]\cr Name of the response.
 #' @param resp_levels \[`character()`]\cr Levels of a categorical response.
 #' @param family \[`dynamitefamily`]\cr Family of the response.
-#' @param eval_type  \[`character(1)`]\cr Either `"predict"` or `"fitted"`.
+#' @param eval_type  \[`character(1)`]\cr Either `"predict"`, `"fitted"` or
+#'   `"loglik"`.
 #' @param has_fixed \[logical(1)]\cr
 #'   Does the channel have time-invariant predictors?
 #' @param has_varying \[logical(1)]\cr
@@ -634,8 +640,9 @@ fitted_categorical <- "
   resp_cols <- c({
     paste0('\"', resp, '_fitted_', resp_levels, '\"', collapse = ', ')
   })
-  maxs <- apply(xbeta, 1, max)
-  mval <- exp(xbeta - (maxs + log(rowSums(exp(xbeta - maxs)))))
+  # maxs <- apply(xbeta, 1, max)
[…]
