### 2022-10-07 8febf14
fix formula output and add match.call output to dynamite
