### 2022-10-11 02b9e70
dont recompile if debug exists, remove () from family
