### 2022-10-11 25f6fa2
log_sum_exp_rows function
