### 2022-10-11 aebbfa8
fix y in loglik computations for loo
