### 2022-10-11 8cc1000
combine channelwise loglikelihoods in loo
