### 2022-10-11 b4bbd78
tweak model_vars so NULL prior and manual default prior leads to same object
