+++ R/default_priors.R
@@ -46,9 +46,9 @@ default_priors <- function(y, channel, mean_gamma, sd_gamma, mean_y, sd_y) {
   if (channel$has_fixed) {
     m <- mean_gamma[channel$J_fixed]
     s <- sd_gamma[channel$J_fixed]
-    channel$beta_prior_npars <- 2L
-    channel$beta_prior_pars <- cbind(m, s, deparse.level = 0)
     channel$beta_prior_distr <- "normal"
+    channel$beta_prior_npars <- 2L
+    channel$beta_prior_pars <- unname(cbind(m, s))
     priors$beta <- data.frame(
       parameter = paste0("beta_", y, "_", names(s)),
       response = y,
@@ -60,9 +60,9 @@ default_priors <- function(y, channel, mean_gamma, sd_gamma, mean_y, sd_y) {
   if (channel$has_varying) {
     m <- mean_gamma[channel$J_varying]
     s <- sd_gamma[channel$J_varying]
-    channel$delta_prior_npars <- 2L
-    channel$delta_prior_pars <- cbind(m, s, deparse.level = 0)
     channel$delta_prior_distr <- "normal"
+    channel$delta_prior_npars <- 2L
+    channel$delta_prior_pars <- unname(cbind(m, s))
     priors$delta <- data.frame(
       parameter = paste0("delta_", y, "_", names(s)),
       response = y,
@@ -70,9 +70,9 @@ default_priors <- function(y, channel, mean_gamma, sd_gamma, mean_y, sd_y) {
       type = "delta",
       category = ""
     )
+    channel$tau_prior_distr <- "normal"
     channel$tau_prior_npars <- 2
     channel$tau_prior_pars <- cbind(0, rep(1, channel$K_varying))
-    channel$tau_prior_distr <- "normal"
     priors$tau <- data.frame(
       parameter = paste0("tau_", y, "_", names(s)),
       response = y,
@@ -101,9 +101,9 @@ default_priors_categorical <- function(y, channel, sd_x, resp_class) {
   if (channel$has_fixed_intercept || channel$has_varying_intercept) {
     m <- rep(0.0, S_y - 1L)
     s <- rep(2.0, S_y - 1L)
-    channel$alpha_prior_npars <- 2L
-    channel$alpha_prior_pars <- cbind(m, s, deparse.level = 0)
     channel$alpha_prior_distr <- "normal"
+    channel$alpha_prior_npars <- 2L
+    channel$alpha_prior_pars <- unname(cbind(m, s))
     priors$alpha <- data.frame(
       parameter = paste0("alpha_", y),
       response = y,
@@ -125,9 +125,9 @@ default_priors_categorical <- function(y, channel, sd_x, resp_class) {
   if (channel$has_fixed) {
     m <- rep(0.0, channel$K_fixed * (S_y - 1L))
     s <- rep(sd_gamma[channel$J_fixed], S_y - 1L)
-    channel$beta_prior_npars <- 2L
     channel$beta_prior_distr <- "normal"
-    channel$beta_prior_pars <- cbind(m, s, deparse.level = 0)
+    channel$beta_prior_npars <- 2L
+    channel$beta_prior_pars <- unname(cbind(m, s))
     priors$beta <- data.frame(
       parameter = paste0("beta_", y, "_", names(s)),
       response = y,
@@ -139,9 +139,9 @@ default_priors_categorical <- function(y, channel, sd_x, resp_class) {
   if (channel$has_varying) {
     m <- rep(0.0, channel$K_varying * (S_y - 1L))
     s <- rep(sd_gamma[channel$J_varying], S_y - 1L)
-    channel$delta_prior_npars <- 2L
-    channel$delta_prior_pars <- cbind(m, s, deparse.level = 0)
[…]
