### 2022-10-12 e23a720
fix cmdstanr backend call and increase optimization level
