### 2022-10-12 b837472
added LFO, update method, cleaned namespace
