### 2022-10-14 55fff49
handle missing data in LFO properly, add print and plot methods for LFO output
