+++ R/lfo.R
@@ -3,27 +3,52 @@
 #' Estimates the leave-future-out (LFO) information criterion for `dynamite`
 #' models using Pareto smoothed importance sampling.
 #'
+#' For multichannel models, the log-likelihoods of all channels are combined.
+#' For models with groups, expected log predictive densities (ELPDs) are
+#' computed independently for each group, but the re-estimation of the model
+#' is triggered if pareto k values of any group exceeds the threshold.
+#'
 #' @export
 #' @export lfo
 #' @aliases lfo
 #' @param x \[`dynamitefit`]\cr The model fit object.
-#' @param L Positive integer defining how many observations should be used for
-#'   the initial fit.
-#' @param verbose If \code{TRUE} (default), print the progress of the LFO
-#'   computations to the console.
-#' @param k_threshold Threshold for the pareto k estimate triggering refit. Default
-#'   is 0.7.
+#' @param L  \[`integer(1)`]\cr Positive integer defining how many time points
+#'   should be used for the initial fit.
+#' @param verbose \[`logical(1)`]\cr If `TRUE` (default), print the progress of
+#'   the LFO computations to the console.
+#' @param k_threshold \[`numeric(1)`]\cr Threshold for the pareto k estimate
+#'   triggering refit. Default is 0.7.
 #' @param ... Additional parameters to `dynamite`.
-#' @return List with components \code{ELPD} (Expected log predictive density),
-#'   \code{ELPDs} (observation-specific ELPDs for each time point),
-#'   \code{ks} (Pareto k values), and \code{refits} (time points where model
-#'   was re-estimated).
+#' @return An `lfo` object which is a `list` with the following components:
+#'   * `ELPD` Expected log predictive density estimate.
+#'   * `ELPD_SE` Standard error of ELPD. This is a crude approximation which
+#'      does not take into account potential serial correlations.
+#'   *  `pareto_k` Pareto k values.
+#'   *  `refits` Time points where model was re-estimated.
+#'   *  `L` L value used in the LFO estimation.
+#'   *  `k_threshold` Threshold used in the LFO estimation.
 #' @examples
 #' \dontrun{
 #' # this gives warnings due to the small number of iterations
-#' lfo(gaussian_example_fit, L = 20)
+#' out <- lfo(gaussian_example_fit, L = 20)
+#' out$ELPD
+#' out$ELPD_SE
 #' }
 lfo <- function(x, L, verbose = TRUE, k_threshold = 0.7, ...) {
+
+  stopifnot_(
+    !is.null(x$stanfit),
+    "No Stan model fit is available."
+  )
+
+  stopifnot_(
+    checkmate::test_flag(x = verbose),
+    "Argument {.arg verbose} must be a single {.cls logical} value."
+  )
+  stopifnot_(
+    checkmate::test_number(x = k_threshold),
+    "Argument {.arg k_threshold} must be a single {.cls numeric} value."
+  )
 
   log_sum_exp <- function(x) {
     max_x <- max(x)
@@ -33,26 +58,32 @@ lfo <- function(x, L, verbose = TRUE, k_threshold = 0.7, ...) {
     log_sum_exp(x) - log(length(x))
   }
 
-  stopifnot_(
-    !is.null(x$stanfit),
-    "No Stan model fit is available."
-  )
   T_ <- x$stan$sampling_vars$T
+
+  stopifnot_(
+    checkmate::test_int(x = L,
+      lower = 0,
[…]
