+++ man/print.lfo.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/lfo.R
+\name{print.lfo}
+\alias{print.lfo}
+\title{Print the results from the LFO}
+\usage{
+\method{print}{lfo}(x, ...)
+}
+\arguments{
+\item{x}{x [\code{lfo}]\cr Output from \code{lfo} function.}
+
+\item{...}{Ignored.}
+}
+\value{
+Returns \code{x} invisibly.
+}
+\description{
+Prints the summary of the leave-future-out cross-validation.
+}
+\examples{
+\dontrun{
+lfo(gaussian_example_fit, L = 20)
+}
+}
