+++ NAMESPACE
@@ -19,9 +19,11 @@ S3method(mcmc_diagnostics,dynamitefit)
 S3method(ndraws,dynamitefit)
 S3method(nobs,dynamitefit)
 S3method(plot,dynamitefit)
+S3method(plot,lfo)
 S3method(predict,dynamitefit)
 S3method(print,dynamitefit)
 S3method(print,dynamiteformula)
+S3method(print,lfo)
 S3method(summary,dynamitefit)
 S3method(update,dynamitefit)
 export(as_draws)
