+++ tests/testthat/test-recovery.R
@@ -118,9 +118,11 @@ test_that("LOO works for AR(1) model", {
   expect_equal(l$estimates,
     structure(c(
       -107.877842970846, 2.86041434691809, 215.755685941693,
-      7.36848739076899, 0.561813071004331, 14.736974781538),
-      dim = 3:2,
-      dimnames = list(c("elpd_loo", "p_loo", "looic"), c("Estimate", "SE"))),
+      7.36848739076899, 0.561813071004331, 14.736974781538
+    ),
+    dim = 3:2,
+    dimnames = list(c("elpd_loo", "p_loo", "looic"), c("Estimate", "SE"))
+    ),
     tolerance = 1
   )
 })
