+++ R/predict.R
@@ -90,7 +90,7 @@
 #' \dontrun{
 #' # Simulate from the prior predictive distribution
 #'
-#' f <-  obs(y ~ lag(y) + varying(~ -1 + x), "gaussian") +
+#' f <- obs(y ~ lag(y) + varying(~ -1 + x), "gaussian") +
 #'   splines(df = 10, noncentered = TRUE)
 #'
 #' # Create data with missing observations
@@ -100,7 +100,8 @@
 #'
 #' # suppress warnings due to the lack of data
 #' suppressWarnings(priors <- get_priors(f,
-#'   data = d, time = "time"))
+#'   data = d, time = "time"
+#' ))
 #'
 #' # modify default priors which can produce exploding behaviour when used
 #' # without data
@@ -109,19 +110,24 @@
 #'   "normal(0.6, 0.1)",
 #'   "normal(-0.2, 0.5)",
 #'   "normal(0.2, 0.1)",
-#'   "normal(0.5, 0.1)")
+#'   "normal(0.5, 0.1)"
+#' )
 #'
 #' # samples from the prior conditional on the first time point and x
-#' fit <- dynamite(f, data = d, time = "time",
-#'   priors = priors, chains = 1)
+#' fit <- dynamite(f,
+#'   data = d, time = "time",
+#'   priors = priors, chains = 1
+#' )
 #'
 #' # simulate new data
 #' pp <- predict(fit)
 #'
 #' library(ggplot2)
 #' ggplot(pp, aes(time, y_new, group = .draw)) +
-#'   geom_line(alpha = 0.1) + theme_bw()
+#'   geom_line(alpha = 0.1) +
+#'   theme_bw()
 #' }
+#'
 predict.dynamitefit <- function(object, newdata = NULL,
                                 type = c("response", "mean", "link"),
                                 funs = list(), impute = c("none", "locf"),
@@ -314,9 +320,9 @@ initialize_predict <- function(object, newdata, type, eval_type, funs, impute,
 #' Obtain Full Individual Level Predictions Or Fitted Values
 #'
 #' @inheritParams initialize_predict
-#' @param simulated A `data.table` containing the simulated values
+#' @param simulated A `data.table` containing the simulated values.
 #' @param observed A `data.table` containing fixed predictors (values
-#'   independent of the posterior draw)
+#'   independent of the posterior draw).
 #' @noRd
 predict_full <- function(object, simulated, observed, type, eval_type,
                          new_levels, n_draws, fixed,
@@ -423,16 +429,6 @@ predict_full <- function(object, simulated, observed, type, eval_type,
 #' @noRd
 predict_summary <- function(object, storage, observed, funs, new_levels,
                             n_draws, fixed, group_var, time_var) {
-  # avoid NSE notes in R CMD check
-  ..group_var <- NULL
-  ..observed <- NULL
-  ..idx_obs <- NULL
-  ..u_time <- NULL
-  ..n_draws <- NULL
-  ..n_time <- NULL
-  ..storage <- NULL
-  ..f <- NULL
-
   formulas_stoch <- get_formulas(object$dformulas$stoch)
   resp_stoch <- get_responses(object$dformulas$stoch)
   lhs_ld <- get_responses(object$dformulas$lag_det)
@@ -457,8 +453,11 @@ predict_summary <- function(object, storage, observed, funs, new_levels,
   idx <- seq.int(n_sim + 1L, 2L * n_sim)
   simulated <- storage[1L, ]
   simulated[, (names(simulated)) := .SD[NA]]
-  simulated <- simulated[rep(1L, 2L * n_sim), ]
[…]
