### 2022-10-24 c070e9b
reduce expand_predict_output memory usage
