### 2022-10-27 459605b
internal data fully using data.table
