+++ R/dynamite.R
@@ -130,22 +130,23 @@
 #'   refresh = 0
 #' )
 #'
-#' library(dplyr)
-#' library(ggplot2)
-#  cf <- coef(fit) |>
-#' group_by(time, variable) |>
-#'   summarise(
-#'     mean = mean(value),
-#'     lwr = quantile(value, 0.025),
-#'     upr = quantile(value, 0.975)
-#'   )
-#'
-#' cf |>
-#'   ggplot(aes(time, mean)) +
-#'   theme_bw() +
-#'   geom_ribbon(aes(ymin = lwr, ymax = upr), alpha = 0.7) +
-#'   geom_line() +
-#'   facet_wrap(~variable, scales = "free_y")
+#' if (requireNamespace("dplyr") && requireNamespace("ggplot2")) {
+#'   library(dplyr)
+#'   library(ggplot2)
+#    cf <- coef(fit) %>%
+#'   group_by(time, variable) %>%
+#'     summarise(
+#'       mean = mean(value),
+#'       lwr = quantile(value, 0.025),
+#'       upr = quantile(value, 0.975)
+#'     )
+#'   cf %>%
+#'     ggplot(aes(time, mean)) +
+#'     theme_bw() +
+#'     geom_ribbon(aes(ymin = lwr, ymax = upr), alpha = 0.7) +
+#'     geom_line() +
+#'     facet_wrap(~variable, scales = "free_y")
+#'   }
 #' }
 #'
 dynamite <- function(dformula, data, group = NULL, time,
@@ -488,6 +489,8 @@ is.dynamitefit <- function(x) {
 #' @noRd
 parse_data <- function(dformula, data, group_var, time_var, verbose) {
   data <- droplevels(data)
+  # droplevels creates a copy already so we can just convert
+  data <- data.table::as.data.table(data)
   data_names <- names(data)
   stopifnot_(
     !is.character(data[[time_var]]),
@@ -505,7 +508,8 @@ parse_data <- function(dformula, data, group_var, time_var, verbose) {
     }
     data[[time_var]] <- as.integer(data[[time_var]])
   }
-  data <- data |> dplyr::mutate(dplyr::across(where(is.character), as.factor))
+  chr_cols <- names(data)[vapply(data, is.character, logical(1L))]
+  data[, (chr_cols) := lapply(.SD, as.factor), .SDcols = chr_cols]
   valid_types <- c("integer", "logical", "double")
   col_types <- vapply(data, typeof, character(1L))
   factor_cols <- vapply(data, is.factor, logical(1L))
@@ -519,17 +523,18 @@ parse_data <- function(dformula, data, group_var, time_var, verbose) {
              {?is/are} not supported."
     )
   )
-  for (i in which(valid_cols & !factor_cols)) {
-    data[, i] <- do.call(
-      paste0("as.", typeof(data[[i]])),
-      args = list(data[[i]])
+  for (j in which(valid_cols & !factor_cols)) {
+    data.table::set(
+      data,
+      j = j,
+      value = do.call(paste0("as.", typeof(data[[j]])), args = list(data[[j]]))
     )
   }
   resp <- get_responses(dformula)
   ordered_factor_resp <- vapply(
     seq_along(resp),
     function(i) {
-      is_categorical(dformula[[i]]$family) && is.ordered(data[, resp[i]])
+      is_categorical(dformula[[i]]$family) && is.ordered(data[[resp[i]]])
     },
     logical(1L)
   )
@@ -543,7 +548,7 @@ parse_data <- function(dformula, data, group_var, time_var, verbose) {
       ))
     }
     for (i in seq_along(rof)) {
-      class(data[, rof[i]]) <- "factor"
[…]

[…]
