### 2022-10-27 79c5b0d
as.data.table method
