+++ R/as_data_table.R
@@ -1,0 +1,426 @@
+#' Extract Samples From a `dynamitefit` Object as a Data Table
+#'
+#' Provides a `data.table` representation of the posterior samples of the model
+#' parameters. See [dynamite::as.data.frame.dynamitefit()] for details.
+#'
+#' @inheritParams as.data.frame.dynamitefit
+#' @export
+#' @examples
+#' as.data.table(
+#'   gaussian_example_fit,
+#'   responses = "y",
+#'   types = "beta",
+#'   summary = FALSE
+#' )
+#'
+as.data.table.dynamitefit <- function(x, row.names = NULL, optional = FALSE,
+                                      responses = NULL, types = NULL,
+                                      summary = TRUE, probs = c(0.05, 0.95),
+                                      include_fixed = TRUE, ...) {
+  stopifnot_(
+    !missing(x),
+    "Argument {.arg x} is missing."
+  )
+  stopifnot_(
+    is.dynamitefit(x),
+    "Argument {.arg x} must be a {.cls dynamitefit} object."
+  )
+  stopifnot_(
+    !is.null(x$stanfit),
+    "No Stan model fit is available."
+  )
+  stopifnot_(
+    checkmate::test_character(
+      x = responses,
+      any.missing = FALSE,
+      min.len = 1L,
+      null.ok = TRUE
+    ),
+    "Argument {.arg responses} must be a {.cls character} vector."
+  )
+  stopifnot_(
+    checkmate::test_character(
+      x = types,
+      any.missing = FALSE,
+      min.len = 1L,
+      null.ok = TRUE
+    ),
+    "Argument {.arg types} must be a {.cls character} vector."
+  )
+  stopifnot_(
+    checkmate::test_flag(x = summary),
+    "Argument {.arg summary} must be a single {.cls logical} value."
+  )
+  stopifnot_(
+    checkmate::test_numeric(
+      x = probs,
+      lower = 0.0,
+      upper = 1.0,
+      any.missing = FALSE,
+      min.len = 1L
+    ),
+    "Argument {.arg probs} must be a {.cls numeric} vector with values between
+     0 and 1."
+  )
+  stopifnot_(
+    checkmate::test_flag(x = include_fixed),
+    "Argument {.arg include_fixed} must be a single {.cls logical} value."
+  )
+  if (is.null(responses)) {
+    responses <- setdiff(unique(x$priors$response), "")
+  } else {
+    z <- responses %in% unique(x$priors$response)
+    stopifnot_(
+      all(z),
+      "Model does not contain response variable{?s} {.var {responses[!z]}}."
+    )
+  }
+  all_types <- c(
+    "alpha", "beta", "delta", "tau", "tau_alpha", "lambda",
+    "sigma_nu", "corr_nu", "sigma", "phi", "nu", "omega", "omega_alpha"
+  )
+  if (is.null(types)) {
+    types <- all_types[seq_len(11L)]
+  } else {
+    types <- onlyif(is.character(types), tolower(types))
+    types <- try(match.arg(types, all_types, TRUE), silent = TRUE)
+    stopifnot_(
+      !inherits(types, "try-error"),
+      "Argument {.arg type} contains unknown types."
+    )
+  }
+  values <- function(type, response) {
+    if (type %in% c("lambda", "corr_nu")) {
+      draws <- rstan::extract(
+        x$stanfit,
+        pars = type,
+        permuted = FALSE
+      )
+    } else {
+      draws <- rstan::extract(
+        x$stanfit,
+        pars = paste0(type, "_", response),
+        permuted = FALSE
+      )
+    }
+    category <- attr(x$stan$responses[[response]], "levels")[-1L]
[…]
