### 2022-10-27 c203f6a
data.tablefied up to loo and lfo
