### 2022-10-28 db8743a
move intercept handling to default model block
