+++ R/stanblocks_families.R
@@ -926,12 +926,14 @@ transformed_parameters_lines_beta <- function(...) {
 
 # Model block -------------------------------------------------------------
 
-model_lines_default <- function(y, idt, noncentered, shrinkage, has_varying,
+model_lines_default <- function(y, idt, obs, noncentered, shrinkage, has_varying,
                                 has_fixed, has_fixed_intercept,
                                 has_varying_intercept, has_random_intercept,
+                                #has_factor,
                                 sigma_nu_prior_distr = "",
                                 alpha_prior_distr = "",
                                 tau_alpha_prior_distr = "",
+                               # lambda_prior_distr = "",
                                 beta_prior_distr = "",
                                 beta_prior_npars = 1L,
                                 delta_prior_distr = "",
@@ -941,6 +943,7 @@ model_lines_default <- function(y, idt, noncentered, shrinkage, has_varying,
                                 K_fixed, K_varying, ...) {
   mtext_u <- "sigma_nu_{y} ~ {sigma_nu_prior_distr};"
   mtext_alpha <- "a_{y} ~ {alpha_prior_distr};"
+  #mtext_lambda <- "lambda_{y} ~ {lambda_prior_distr};"
 
   if (noncentered) {
     mtext_omega <- "omega_raw_alpha_{y} ~ std_normal();"
@@ -1036,7 +1039,33 @@ model_lines_default <- function(y, idt, noncentered, shrinkage, has_varying,
     mtext_tau <- "tau_{y}[{{{cs(1:K_varying)}}}] ~ {tau_prior_distr};"
   }
 
-  paste_rows(
+  intercept_alpha <- ifelse_(
+    has_fixed_intercept,
+    glue::glue("alpha_{y}"),
+    ifelse_(
+      has_varying_intercept,
+      glue::glue("alpha_{y}[t]"),
+      ""
+    )
+  )
+  intercept_nu <- ifelse_(
+    has_random_intercept,
+    ifelse(
+      nzchar(obs),
+      glue::glue("nu_{y}[{obs}]"),
+      glue::glue("nu_{y}")
+    ),
+    ""
+  )
+  plus <- ifelse_(nzchar(intercept_alpha) && has_random_intercept, " + ", "")
+  intercept <- ifelse_(
+    nzchar(intercept_alpha) || has_random_intercept,
+    glue::glue("{intercept_alpha}{plus}{intercept_nu}"),
+    "0"
+  )
+
+  list(text = paste_rows(
+    #onlyif(has_factor, mtext_lambda),
     onlyif(has_random_intercept, mtext_u),
     onlyif(has_fixed_intercept, mtext_fixed_intercept),
     onlyif(has_varying_intercept, mtext_varying_intercept),
@@ -1044,7 +1073,7 @@ model_lines_default <- function(y, idt, noncentered, shrinkage, has_varying,
     onlyif(has_varying, mtext_varying),
     onlyif(has_varying, mtext_tau),
     .indent = idt(c(1, 1, 1, 1, 1, 1))
-  )
+  ), intercept = intercept)
 }
 
 model_lines_categorical <- function(y, idt, obs, noncentered, shrinkage,
@@ -1257,30 +1286,7 @@ model_lines_gaussian <- function(y, idt, obs, has_fixed, has_varying,
[…]
