+++ R/lfo.R
@@ -92,46 +92,60 @@ lfo <- function(x, L, verbose = TRUE, k_threshold = 0.7, ...) {
     new_levels = "none",
     global_fixed = FALSE,
     n_draws = NULL,
-    expand = FALSE
+    expand = FALSE,
+    df = FALSE
   )$simulated
 
   n_draws <- ndraws(x)
   # sum the log-likelihood over the channels and non-missing time points
   # for each group, time, and draw
   # drop those id&time pairs which contain NA
-  lls <- out |>
-    dplyr::filter(.data[[time]] > timepoints[L]) |>
-    dplyr::mutate(loglik = rowSums(dplyr::across(
-      dplyr::ends_with("_loglik")
-    ))) |>
-    dplyr::select(-dplyr::ends_with("_loglik")) |>
-    tidyr::drop_na()
+  lls <- na.omit(out[
+    time > timepoints[L], ,
+    env = list(time = time, timepoints = timepoints, L = L)
+  ][,
+    loglik := rowSums(.SD),
+    .SDcols = patterns("_loglik$")
+  ][,
+    .SD,
+    .SDcols = !patterns("_loglik$")
+  ])
 
   elpds <- vector("list", T_ - L)
-  elpds[[1L]] <- lls |>
-    dplyr::filter(.data[[time]] == timepoints[L + 1L]) |>
-    dplyr::group_by(.data[[time]], onlyif(!is.null(id), .data[[id]])) |>
-    dplyr::summarise(elpd = log_mean_exp(.data$loglik), .groups = "keep") |>
-    dplyr::pull("elpd")
+  elpds[[1L]] <- lls[
+    time == timepoints[L + 1L], ,
+    env = list(time = time, timepoints = timepoints, L = L)
+  ][,
+    elpd := log_mean_exp(loglik),
+    by = list(time, id),
+    env = list(log_mean_exp = "log_mean_exp")
+  ][["elpd"]]
 
   i_refit <- L
   refits <- timepoints[L]
   ks <- vector("list", T_ - L - 1L)
 
   for (i in seq.int(L + 1L, T_ - 1L)) {
-    if (nrow(lls |> dplyr::filter(.data[[time]] == i + 1L)) > 0L) {
-      logratio <- lls |>
-        dplyr::filter(
-          .data[[time]] > timepoints[i_refit] &
-            .data[[time]] <= timepoints[i]
-        ) |>
-        dplyr::group_by(onlyif(!is.null(id), .data[[id]]), .data$.draw) |>
-        dplyr::summarise(logratio = sum(.data$loglik), .groups = "keep")
+    if (lls[time == i + 1L, .N, env = list(time = time, i = i)] > 0L) {
+      logratio <- lls[
+        time > timepoints[i_refit] & time <= timepoints[i], ,
+        env = list(
+          time = time,
+          timepoints = timepoints,
+          i = i,
+          i_refit = i_refit
+        )
+      ][,
+        logratio := sum(loglik),
+        by = list(id, .draw)
+      ]
 
-      psis_obj <- suppressWarnings(loo::psis(
-        matrix(logratio$logratio, nrow = n_draws),
-        r_eff = NA
-      ))
+      psis_obj <- suppressWarnings(
+        loo::psis(
+          matrix(logratio$logratio, nrow = n_draws),
+          r_eff = NA
+        )
+      )
       k <- loo::pareto_k_values(psis_obj)
       ks[[i - L]] <- k
       if (any(k > k_threshold)) {
@@ -143,7 +157,7 @@ lfo <- function(x, L, verbose = TRUE, k_threshold = 0.7, ...) {
         refits <- c(refits, timepoints[i])
         d <- data.table::copy(x$data)
         set_na <- d[[x$time_var]] > timepoints[i]
-        d[set_na, (responses) := NA]
[…]
