+++ R/stanblocks_families.R
@@ -518,7 +518,7 @@ transformed_parameters_lines_default <- function(y, idt, noncentered, shrinkage,
                                                  J, K, K_fixed, K_varying,
                                                  L_fixed, L_varying) {
   if (noncentered) {
-    lambda_term <- ifelse_(shrinkage, " * lambda[i - 1];", ";")
+    xi_term <- ifelse_(shrinkage, " * xi[i - 1];", ";")
     declare_omega <- paste_rows(
       "// Spline coefficients",
       "matrix[{K_varying}, D] omega_{y};",
@@ -530,7 +530,7 @@ transformed_parameters_lines_default <- function(y, idt, noncentered, shrinkage,
       "for (i in 2:D) {{",
       paste0(
         "omega_{y}[, i] = omega_{y}[, i - 1] + ",
-        "omega_raw_{y}[, i] .* tau_{y}{lambda_term}"
+        "omega_raw_{y}[, i] .* tau_{y}{xi_term}"
       ),
       "}}",
       .indent = idt(c(1, 1, 2, 1)),
@@ -607,7 +607,7 @@ transformed_parameters_lines_default <- function(y, idt, noncentered, shrinkage,
     state_fixed_intercept <- character(0L)
   }
   if (noncentered) {
-    lambda_term <- ifelse_(shrinkage, " * lambda[i - 1];", ";")
+    xi_term <- ifelse_(shrinkage, " * xi[i - 1];", ";")
     declare_omega_alpha <- paste_rows(
       "// Spline coefficients",
       "row_vector[D] omega_alpha_{y};",
@@ -619,7 +619,7 @@ transformed_parameters_lines_default <- function(y, idt, noncentered, shrinkage,
       "for (i in 2:D) {{",
       paste0(
         "omega_alpha_{y}[i] = omega_alpha_{y}[i - 1] + ",
-        "omega_raw_alpha_{y}[i - 1] * tau_alpha_{y}{lambda_term}"
+        "omega_raw_alpha_{y}[i - 1] * tau_alpha_{y}{xi_term}"
       ),
       "}}",
       .indent = idt(c(1, 1, 2, 1)),
@@ -682,7 +682,7 @@ transformed_parameters_lines_categorical <- function(y, idt, noncentered,
                                                      L_fixed, L_varying,
                                                      S, ...) {
   if (noncentered) {
-    lambda_term <- ifelse_(shrinkage, " * lambda[i - 1];", ";")
+    xi_term <- ifelse_(shrinkage, " * xi[i - 1];", ";")
     declare_omega <- paste_rows(
       "// Spline coefficients",
       "matrix[{K_varying}, D] omega_{y}[{S - 1}];",
@@ -695,7 +695,7 @@ transformed_parameters_lines_categorical <- function(y, idt, noncentered,
       "for (i in 2:D) {{",
       paste0(
         "omega_{y}[s, , i] = omega_{y}[s, , i - 1] + ",
-        "omega_raw_{y}[s, , i] .* tau_{y}{lambda_term}"
+        "omega_raw_{y}[s, , i] .* tau_{y}{xi_term}"
       ),
       "}}",
       "}}",
@@ -796,7 +796,7 @@ transformed_parameters_lines_categorical <- function(y, idt, noncentered,
     )
   }
   if (noncentered) {
-    lambda_term <- ifelse_(shrinkage, " * lambda[i - 1];", ";")
+    xi_term <- ifelse_(shrinkage, " * xi[i - 1];", ";")
     declare_omega_alpha <- paste_rows(
       "// Spline coefficients",
[…]
