### 2022-10-28 b976fbb
lambda to xi in tests
