+++ man/fitted.dynamitefit.Rd
@@ -4,7 +4,7 @@
 \alias{fitted.dynamitefit}
 \title{Extract Fitted Values of a Dynamite Model}
 \usage{
-\method{fitted}{dynamitefit}(object, newdata = NULL, n_draws = NULL, expand = TRUE, ...)
+\method{fitted}{dynamitefit}(object, newdata = NULL, n_draws = NULL, expand = TRUE, df = TRUE, ...)
 }
 \arguments{
 \item{object}{[\code{dynamitefit}]\cr The model fit object.}
@@ -26,6 +26,9 @@ predicted values, and the second contains the original \code{newdata}.
 Setting \code{expand} to \code{FALSE} can help conserve memory because \code{newdata}
 is not replicated \code{n_draws} times in the output.
 This argument is ignored if \code{funs} are provided.}
+
+\item{df}{[\code{logical(1)}]\cr If \code{TRUE} (default) the output
+consists of \code{data.frame} objects, and \code{data.table} objects otherwise.}
 
 \item{...}{Ignored.}
 }
