### 2022-11-04 9a54565
add grouping variable if group is NULL is dynamite
