+++ R/predict_helpers.R
@@ -62,25 +62,23 @@ check_newdata <- function(object, newdata) {
 parse_newdata <- function(dformula, newdata, data, type, eval_type,
                           families_stoch, resp_stoch, categories,
                           clear_names, new_levels, group_var, time_var) {
-  if (!is.null(group_var)) {
-    stopifnot_(
-      group_var %in% names(newdata),
-      "Can't find grouping variable {.var {group_var}} in {.var newdata}."
-    )
-    group <- newdata[[group_var]]
-    group <- unique(group)
-    extra_levels <- unique(group[!group %in% data[[group_var]]])
-    stopifnot_(
-      all(group %in% data[[group_var]]) || !identical(new_levels, "none"),
-      c(
-        "Grouping variable {.var {group_var}} contains unknown levels:",
-        `x` = "Level{?s} {.val {as.character(extra_levels)}}
-               {?is/are} not present in the original data.",
-        `i` = "{.strong Note:} argument {.var new_levels} is {.val none}
-               which disallows new levels."
-      )
-    )
+  if (!group_var %in% names(newdata)) {
+    orig <- sort(data[[group_var]])[1L]
+    data.table::set(newdata, j = group_var, value = orig)
   }
+  group <- newdata[[group_var]]
+  group <- unique(group)
+  extra_levels <- unique(group[!group %in% data[[group_var]]])
+  stopifnot_(
+    all(group %in% data[[group_var]]) || !identical(new_levels, "none"),
+    c(
+      "Grouping variable {.var {group_var}} contains unknown levels:",
+      `x` = "Level{?s} {.val {as.character(extra_levels)}}
+             {?is/are} not present in the original data.",
+      `i` = "{.strong Note:} argument {.var new_levels} is {.val none}
+             which disallows new levels."
+    )
+  )
   stopifnot_(
     time_var %in% names(newdata),
     "Can't find time index variable {.var {time_var}} in {.var newdata}."
@@ -217,84 +215,57 @@ parse_funs <- function(object, funs) {
 #' @inheritParams dynamite
 #' @noRd
 fill_time_predict <- function(data, group_var, time_var, time_scale) {
-  has_groups <- !is.null(group_var)
-  if (has_groups) {
-    time_duplicated <- data[,
-      any(duplicated(time_var)),
-      by = group_var,
-      env = list(time_var = time_var)
-    ]$V1
-    d <- which(time_duplicated)
-    stopifnot_(
-      all(!time_duplicated),
-      c(
-        "Each time index must correspond to a single observation per group:",
-        `x` = "{cli::qty(d)}Group{?s} {.var {d}} of {.var {group_var}}
-               {cli::qty(d)}{?has/have} duplicate observations."
-      )
+  time_duplicated <- data[,
+    any(duplicated(time_var)),
+    by = group_var,
+    env = list(time_var = time_var)
+  ]$V1
+  d <- which(time_duplicated)
+  stopifnot_(
+    all(!time_duplicated),
+    c(
+      "Each time index must correspond to a single observation per group:",
+      `x` = "{cli::qty(d)}Group{?s} {.var {d}} of {.var {group_var}}
+             {cli::qty(d)}{?has/have} duplicate observations."
     )
-  } else {
-    stopifnot_(
-      all(!duplicated(data[[time_var]])),
-      "Each time index must correspond to a single observation."
-    )
-  }
+  )
   time <- sort(unique(data[[time_var]]))
   if (length(time) > 1L) {
     original_order <- colnames(data)
     full_time <- seq(time[1L], time[length(time)], by = time_scale)
-    if (has_groups) {
-      time_groups <- data[,
-        {
-          has_missing = !identical(time_var, full_time)
[…]
