+++ tests/testthat/test-edgecases.R
@@ -383,7 +383,6 @@ test_that("data expansion to full time scale works", {
   expect_equal(fit$data, expected_data, ignore_attr = TRUE)
   # no groups
   test_data_single <- test_data[test_data[["group"]] == 1L, ]
-  test_data_single$group <- NULL
   mis_rows_single <- c(3, 5, 6, 9)
   test_data_single_mis <- test_data_single[-mis_rows_single, ]
   fit_single <- dynamite(
@@ -397,7 +396,7 @@ test_that("data expansion to full time scale works", {
     debug = debug
   )
   expected_data_single <- test_data_single
-  expected_data_single[mis_rows_single, seq(2, ncol(test_data_single))] <- NA
+  expected_data_single[mis_rows_single, seq(3, ncol(test_data_single))] <- NA
   expected_data_single$x1 <- factor(expected_data_single$x1)
   expected_data_single <- droplevels(expected_data_single)
   expected_data_single$trials <- NULL
