### 2022-11-18 58383b5
modelling latent factors work, not fully tested
