+++ R/stanblocks_families.R
@@ -41,7 +41,7 @@ vectorizable_prior <- function(x) {
 #' @param has_random_intercept \[`logical(1)`]\cr Does the channel have a
 #'   random intercept?
 #' @param noncentered \[`logical(1)`]\cr Should the noncentered parametrization
-#'   be used?
+#'   be used for splines?
 #' @param shrinkage \[`logical(1)`]\cr Should the common global shrinkage
 #'   parameter be used?
 #' @param lb \[`double(1)`]\cr Lower bound for the `tau` parameter.
@@ -377,7 +377,10 @@ parameters_lines_default <- function(y, idt, noncentered, lb, has_fixed,
                                      has_varying, has_fixed_intercept,
                                      has_varying_intercept,
                                      has_random_intercept,
-                                     K_fixed, K_varying) {
+                                     K_fixed, K_varying,
+                                     has_lfactor, noncentered_psi,
+                                     noncentered_lambda,
+                                     nonzero_lambda) {
   oname <- ifelse_(noncentered, "omega_raw_", "omega_")
   paste_rows(
     onlyif(
@@ -405,6 +408,22 @@ parameters_lines_default <- function(y, idt, noncentered, lb, has_fixed,
     onlyif(
       has_varying_intercept,
       "real<lower={lb}> tau_alpha_{y}; // SD for the random walk"
+    ),
+    onlyif(
+      has_lfactor && nonzero_lambda,
+      "real<lower=0> tau_psi_{y}; // SD for for the random walk"
+    ),
+    onlyif(
+      has_lfactor,
+      "real<lower=0> sigma_lambda_{y}; // SD of factor loadings"
+    ),
+    onlyif(
+      has_lfactor,
+      "vector[N - 1] lambda_raw_{y}; // raw factor loadings"
+    ),
+    onlyif(
+      has_lfactor  && noncentered_psi,
+      "real omega_raw_psi_1_{y}; // factor spline coef for first time point"
     ),
     .indent = idt(1)
   )
@@ -516,7 +535,11 @@ transformed_parameters_lines_default <- function(y, idt, noncentered, shrinkage,
                                                  has_fixed_intercept,
                                                  has_varying_intercept,
                                                  J, K, K_fixed, K_varying,
-                                                 L_fixed, L_varying) {
+                                                 L_fixed, L_varying,
+                                                 has_lfactor,
+                                                 noncentered_psi,
+                                                 noncentered_lambda,
+                                                 nonzero_lambda) {
   if (noncentered) {
     xi_term <- ifelse_(shrinkage, " * xi[i - 1];", ";")
     declare_omega <- paste_rows(
@@ -539,7 +562,7 @@ transformed_parameters_lines_default <- function(y, idt, noncentered, shrinkage,
   }
 
   declare_delta <- paste_rows(
-    "// Fixed coefficients",
+    "// Varying coefficients",
     "vector[{K_varying}] delta_{y}[T];",
     .indent = idt(c(0, 1)),
     .parse = FALSE
@@ -655,19 +678,67 @@ transformed_parameters_lines_default <- function(y, idt, noncentered, shrinkage,
     .parse = FALSE
   )
 
+  m <- ifelse_(nonzero_lambda, "1 + ", "")
+  if (noncentered_lambda) {
+    declare_lambda <- paste_rows(
+      "// hard sum-to-zero constraint",
+      "vector[N] lambda_std_{y} = A_qr * lambda_raw_{y};",
+      "vector[N] lambda_{y} = {m}sigma_lambda_{y} * lambda_std_{y};",
[…]
