+++ R/dynamiteformula.R
@@ -431,6 +431,8 @@ add_dynamiteformula <- function(e1, e2) {
     out <- set_splines(e1, e2)
   } else if (inherits(e2, "random")) {
     out <- set_random(e1, e2)
+  } else if (inherits(e2, "latent_factor")) {
+    out <- set_lfactor(e1, e2)
   } else {
     stop_(
       "Unable to add an object of class {.cls {class(e2)}}
@@ -541,5 +543,19 @@ set_random <- function(e1, e2) {
     "Multiple definitions for random intercepts."
   )
   attr(e1, "random") <- e2
+  e1
+}
+
+#' Set the Latent Factors of the Model
+#'
+#' @param e1 A `dynamiteformula` object.
+#' @param e2 A `latent_factor` object.
+#' @noRd
+set_lfactor <- function(e1, e2) {
+  stopifnot_(
+    is.null(attr(e1, "lfactor")) || attr(e2, "lfactor"),
+    "Multiple definitions for latent factors."
+  )
+  attr(e1, "lfactor") <- e2
   e1
 }
