+++ R/stanblocks.R
@@ -44,7 +44,8 @@ create_functions <- function(dformula, idt, vars) {
 #' @noRd
 create_data <- function(dformula, idt, vars) {
   has_splines <- any(unlist(lapply(vars, "[[", "has_varying"))) ||
-    any(unlist(lapply(vars, "[[", "has_varying_intercept")))
+    any(unlist(lapply(vars, "[[", "has_varying_intercept"))) ||
+    any(unlist(lapply(vars, "[[", "has_lfactor")))
   mtext <- paste_rows(
     "int<lower=1> T; // number of time points",
     "int<lower=1> N; // number of individuals",
@@ -54,6 +55,7 @@ create_data <- function(dformula, idt, vars) {
     onlyif(has_splines, "int<lower=1> D; // number of B-splines"),
     onlyif(has_splines, "matrix[D, T] Bs; // B-spline basis matrix"),
     "int<lower=0> M; // number of channels with random intercept",
+    "int<lower=0> P; // number of channels with latent factor",
     .indent = idt(1),
     .parse = FALSE
   )
@@ -82,12 +84,28 @@ create_transformed_data <- function(dformula, idt, vars) {
     declarations[i] <- tr_data$declarations
     statements[i] <- tr_data$statements
   }
+
+  declare_AQR <- "matrix[N, N - 1] A_qr;"
+  state_AQR <- paste_rows(
+    "{",
+    "matrix[N, N] A = diag_matrix(rep_vector(1, N));",
+    "for (i in 1:K - 1) A[K, i] = -1;",
+    "A[K, K] = 0;",
+    "A_qr = qr_Q(A)[ , 1:(N - 1)];",
+    "}",
+    .indent = idt(c(0, 2, 2, 2, 2, 1)),
+    .parse = FALSE
+  )
+  has_lfactor <- length(attr(dformula, "lfactor")$responses) > 0L
   paste_rows(
     "transformed data {",
-    "  // Parameters for vectorized priors",
+    "// Parameters for vectorized priors",
     declarations,
+    onlyif(has_lfactor, declare_AQR),
     statements,
+    onlyif(has_lfactor, state_AQR),
     "}",
+    .indent = idt(c(0, 1, 0, 1, 0, 1, 0)),
     .parse = FALSE
   )
 }
@@ -124,19 +142,39 @@ create_parameters <- function(dformula, idt, vars) {
       .indent = idt(c(1, 1, 1))
     )
   )
+
+  lfactortext <- ifelse_(
+    identical(length(attr(dformula, "lfactor")$responses), 0L),
+    "",
+    paste_rows(
+      "// Latent factor splines",
+      onlyif(
+        attr(dformula, "lfactor")$correlated,
+        "cholesky_factor_corr[P] L_lf; // Cholesky for correlated factors"
+      ),
+      ifelse_(
+        attr(dformula, "lfactor")$noncentered_psi,
+        "matrix[P, D - 1] omega_raw_psi;", "matrix[P, D] omega_raw_psi;"
+      ),
+      .indent = idt(c(1, 1, 1))
+    )
+  )
+
   pars <- character(length(dformula))
   for (i in seq_along(dformula)) {
     family <- dformula[[i]]$family$name
     line_args <- c(list(y = vars[[i]]$resp, idt = idt), vars[[i]])
     pars[i] <- lines_wrap("parameters", family, line_args)
   }
-  paste_rows("parameters {", splinetext, randomtext, pars, "}", .parse = FALSE)
+  paste_rows("parameters {", splinetext, randomtext, lfactortext, pars, "}",
+    .parse = FALSE)
 }
 
 #' @describeIn create_function Create the 'Transformed Parameters'
 #'   Block of the Stan Model Code
 #' @noRd
[…]
