### 2022-11-19 f1709d9
fix QR matrix
