### 2022-11-19 5d9c47f
as.data.frame and plots for latent factors
