### 2022-12-01 34592b4
enable type argument when using funs in predict
