### 2022-12-02 98b6104
fix yet another issue with time indexing in lfo
