### 2022-12-08 0f8b7c6
update README, data.table dev and remove R 4.1 dep
