+++ README.Rmd
@@ -49,24 +49,9 @@ grant 331817 ([PREDLIFE](https://sites.utu.fi/predlife/en/)).
 
 ## Installation
 
-`dynamite` uses R's native pipe operator, so R version 4.1.0 or newer is required.
-
-<!-- `dynamite` also relies on some of the features of [Stan](https://mc-stan.org/)'s  -->
-<!-- which are not yet available on the [CRAN](https://cran.r-project.org/) version  -->
-<!-- of [rstan](https://CRAN.R-project.org/package=rstan) package. Therefore you  -->
-<!-- need to install the latest `rstan` and `StanHeaders` packages: -->
-<!-- ```{r, eval = FALSE} -->
-<!-- # run the next line if you already have rstan installed: -->
-<!-- # remove.packages(c("StanHeaders", "rstan")) -->
-<!-- install.packages("rstan", repos = c("https://mc-stan.org/r-packages/",  -->
-<!--   getOption("repos"))) -->
-<!-- ``` -->
-<!-- (See more help on [RStan wiki](https://github.com/stan-dev/rstan/wiki/RStan-Getting-Started)) -->
-
-Also, version `1.14.3` or newer of the 
-[data.table](https://github.com/Rdatatable/data.table) package is required, 
-which can be installed via:
-
+`dynamite` requires the development version of the 
+[data.table](https://github.com/Rdatatable/data.table) package 
+(version `1.14.3` or newer), which can be installed via:
 ```{r, eval = FALSE}
 # run the next line if you do not have any version of the data.table installed:
 # install.packages("data.table")
@@ -75,7 +60,7 @@ which can be installed via:
 data.table::update.dev.pkg()
 ```
 
-After these steps, you can install the development version of dynamite from 
+After these steps, you can install the development version of `dynamite` from 
 [GitHub](https://github.com/) with:
 
 ```{r, eval = FALSE}
