### 2022-12-08 bfb1b69
as.data.table method, replace aes_string, fix examples, update gaussian_example_fit run_extended
