+++ R/fitted.R
@@ -27,11 +27,13 @@
 #'   refresh = 0
 #' )
 #'
-#  if (requireNamespace("dplyr")) {
+#' if (requireNamespace("dplyr") &&
+#'     requireNamespace("tidyr") &&
+#'     base::getRversion() >= "4.1.0") {
 #'   # One-step ahead samples (fitted values) from the posterior
 #'   # (first time point is fixed due to lag in the model):
-#'   fitted(fit) %>%
-#'     dplyr::filter(time > 2) %>%
+#'   fitted(fit) |>
+#'     dplyr::filter(time > 2) |>
 #'     ggplot2::ggplot(aes(time, LakeHuron_fitted, group = .draw)) +
 #'     ggplot2::geom_line(alpha = 0.5) +
 #'     # observed values
@@ -39,13 +41,14 @@
 #'     ggplot2::theme_bw()
 #'
 #'   # Posterior predictive distribution given the first time point:
-#'   predict(fit, type = "mean") %>%
-#'     dplyr::filter(time > 2) %>%
+#'   predict(fit, type = "mean") |>
+#'     dplyr::filter(time > 2) |>
 #'     ggplot2::ggplot(aes(time, LakeHuron_mean, group = .draw)) +
 #'     ggplot2::geom_line(alpha = 0.5) +
 #'     # observed values
 #'     ggplot2::geom_line(aes(y = LakeHuron), colour = "tomato") +
 #'     ggplot2::theme_bw()
+#' }
 #' }
 #'
 fitted.dynamitefit <- function(object, newdata = NULL, n_draws = NULL,
