+++ R/dynamite.R
@@ -134,21 +134,23 @@
 #'   refresh = 0
 #' )
 #'
-#' if (requireNamespace("dplyr") && requireNamespace("ggplot2")) {
-#    cf <- coef(fit) %>%
-#'   dplyr::group_by(time, variable) %>%
+#' if (requireNamespace("dplyr") &&
+#'     requireNamespace("tidyr") &&
+#'     base::getRversion() >= "4.1.0") {
+#    cf <- coef(fit) |>
+#'   dplyr::group_by(time, variable) |>
 #'     dplyr::summarise(
 #'       mean = mean(value),
 #'       lwr = quantile(value, 0.025),
 #'       upr = quantile(value, 0.975)
 #'     )
-#'   cf %>%
+#'   cf |>
 #'     ggplot2::ggplot(aes(time, mean)) +
 #'     ggplot2::theme_bw() +
 #'     ggplot2::geom_ribbon(aes(ymin = lwr, ymax = upr), alpha = 0.7) +
 #'     ggplot2::geom_line() +
 #'     ggplot2::facet_wrap(~variable, scales = "free_y")
-#'   }
+#' }
 #' }
 #'
 dynamite <- function(dformula, data, group = NULL, time,

[…]
