+++ README.md
@@ -21,38 +21,27 @@ series) data comprising of multiple measurements per multiple
 individuals measured in time. The main features distinguishing the
 package and the underlying methodology from many other approaches are:
 
--   Support for both time-invariant and time-varying effects modeled via
-    B-splines.
--   Joint modeling of multiple measurements per individual (multiple
-    channels) based directly on the assumed data generating process.
--   Support for non-Gaussian observations: Currently Gaussian,
-    Categorical, Poisson, Bernoulli, Binomial, Negative Binomial, Gamma,
-    Exponential, and Beta distributions are available and these can be
-    mixed arbitrarily in multichannel models.
--   Allows evaluating realistic long-term counterfactual predictions
-    which take into account the dynamic structure of the model by
-    posterior predictive distribution simulation.
--   Transparent quantification of parameter and predictive uncertainty
-    due to a fully Bayesian approach.
--   User-friendly and efficient R interface with state-of-the-art
-    estimation via Stan.
+- Support for both time-invariant and time-varying effects modeled via
+  B-splines.
+- Joint modeling of multiple measurements per individual (multiple
+  channels) based directly on the assumed data generating process.
+- Support for non-Gaussian observations: Currently Gaussian,
+  Categorical, Poisson, Bernoulli, Binomial, Negative Binomial, Gamma,
+  Exponential, and Beta distributions are available and these can be
+  mixed arbitrarily in multichannel models.
+- Allows evaluating realistic long-term counterfactual predictions which
+  take into account the dynamic structure of the model by posterior
+  predictive distribution simulation.
+- Transparent quantification of parameter and predictive uncertainty due
+  to a fully Bayesian approach.
+- User-friendly and efficient R interface with state-of-the-art
+  estimation via Stan. Both `rstan` and `cmdstanr` backends are
+  supported.
 
 The `dynamite` package is developed with the support of Academy of
 Finland grant 331817 ([PREDLIFE](https://sites.utu.fi/predlife/en/)).
 
 ## Installation
-
-<!-- `dynamite` also relies on some of the features of [Stan](https://mc-stan.org/)'s  -->
-<!-- which are not yet available on the [CRAN](https://cran.r-project.org/) version  -->
-<!-- of [rstan](https://CRAN.R-project.org/package=rstan) package. Therefore you  -->
-<!-- need to install the latest `rstan` and `StanHeaders` packages: -->
-<!-- ```{r, eval = FALSE} -->
-<!-- # run the next line if you already have rstan installed: -->
-<!-- # remove.packages(c("StanHeaders", "rstan")) -->
-<!-- install.packages("rstan", repos = c("https://mc-stan.org/r-packages/",  -->
-<!--   getOption("repos"))) -->
-<!-- ``` -->
-<!-- (See more help on [RStan wiki](https://github.com/stan-dev/rstan/wiki/RStan-Getting-Started)) -->
 
 `dynamite` requires the development version of the
 [data.table](https://github.com/Rdatatable/data.table) package (version
@@ -63,7 +52,7 @@ Finland grant 331817 ([PREDLIFE](https://sites.utu.fi/predlife/en/)).
 # install.packages("data.table")
 
 # install the latest development version:
-data.table::update.dev.pkg()
+data.table::update_dev_pkg()
 ```
 
 After these steps, you can install the development version of `dynamite`
@@ -147,28 +136,29 @@ For more examples, see the package vignette.
 
 ## Related packages
 
--   The `dynamite` package uses Stan via
-    [`rstan`](https://CRAN.R-project.org/package=rstan) (see also
-    <https://mc-stan.org>), which is a probabilistic programming
-    language for general Bayesian modelling.
[…]
