### 2022-12-09 b4eb682
document latent factor and remove todos, draw to .draw in predict summary
