### 2022-12-09 b3513af
cleanup  lfactor preparation
