### 2022-12-12 aca2d69
fix lfactor construction, remove extra intercept
