### 2022-12-12 fb0f4ba
check for windows in update example
