+++ R/update.R
@@ -20,7 +20,10 @@
 #' @examples
 #' \dontrun{
 #' # re-estimate the example fit without thinning:
-#' fit <- update(gaussian_example_fit, thin = 1)
+#' # As the model is compiled in windows, this will fail in other platforms
+#' if (.Platform$OS.type == "windows") {
+#'   fit <- update(gaussian_example_fit, thin = 1)
+#' }
 #' }
 #'
 update.dynamitefit <- function(object, dformula = NULL, data = NULL,
