### 2022-12-12 bb618c6
latent factor test model, fix latent factor constructions
