+++ R/stanblocks.R
@@ -96,7 +96,7 @@ create_transformed_data <- function(dformula, idt, vars) {
     .indent = idt(c(0, 2, 2, 2, 2, 1)),
     .parse = FALSE
   )
-  has_lfactor <- length(attr(dformula, "lfactor")$responses) > 0L
+  has_lfactor <- any(unlist(lapply(vars, "[[", "has_lfactor")))
   paste_rows(
     "transformed data {",
     "// Parameters for vectorized priors",
@@ -126,17 +126,18 @@ create_parameters <- function(dformula, idt, vars) {
       .indent = idt(1)
     )
   )
+
   randomtext <- ifelse_(
-    identical(length(attr(dformula, "random")$responses), 0L),
+    identical(length(vars$random_defs$responses), 0L),
     "",
     paste_rows(
       "// Random intercepts",
       onlyif(
-        attr(dformula, "random")$correlated,
+        vars$random_defs$correlated,
         "cholesky_factor_corr[M] L; // Cholesky for correlated intercepts"
       ),
       ifelse_(
-        attr(dformula, "random")$noncentered,
+        vars$random_defs$noncentered,
         "matrix[N, M] nu_raw;", "vector[M] nu_raw[N];"
       ),
       .indent = idt(c(1, 1, 1))
@@ -144,16 +145,16 @@ create_parameters <- function(dformula, idt, vars) {
   )
 
   lfactortext <- ifelse_(
-    identical(length(attr(dformula, "lfactor")$responses), 0L),
+    identical(length(vars$lfactor_defs$responses), 0L),
     "",
     paste_rows(
       "// Latent factor splines",
       onlyif(
-        attr(dformula, "lfactor")$correlated,
+        vars$lfactor_defs$correlated,
         "cholesky_factor_corr[P] L_lf; // Cholesky for correlated factors"
       ),
       ifelse_(
-        attr(dformula, "lfactor")$noncentered_psi,
+        vars$lfactor_defs$noncentered_psi,
         "matrix[P, D - 1] omega_raw_psi;", "matrix[P, D] omega_raw_psi;"
       ),
       .indent = idt(c(1, 1, 1))
@@ -176,12 +177,12 @@ create_parameters <- function(dformula, idt, vars) {
 create_transformed_parameters <- function(dformula, idt, vars) {
 
   randomtext <- ""
-  nus <- attr(dformula, "random")$responses
+  nus <- vars$random_defs$responses
   M <- length(nus)
   if (M > 0) {
-    if (attr(dformula, "random")$noncentered) {
+    if (vars$random_defs$noncentered) {
       randomtext <- ifelse_(
-        attr(dformula, "random")$correlated,
+        vars$random_defs$correlated,
         paste_rows(
           "matrix[N, M] nu = nu_raw * L';",
           glue::glue("vector[N] nu_{nus} = sigma_nu_{nus} * nu[, {1:M}];"),
@@ -203,15 +204,15 @@ create_transformed_parameters <- function(dformula, idt, vars) {
 
 
   lfactortext <- ""
-  psis <- attr(dformula, "lfactor")$responses
+  psis <- vars$lfactor_defs$responses
   P <- length(psis)
   if (P > 0) {
-    if (attr(dformula, "lfactor")$noncentered_psi) {
-      tau_psi <- ifelse(attr(dformula, "lfactor")$nonzero_lambda,
+    if (vars$lfactor_defs$noncentered_psi) {
+      tau_psi <- ifelse(vars$lfactor_defs$nonzero_lambda,
         paste0("tau_psi_", psis, " * "), "")
       # create noise terms for latent factors
       lfactortext <- ifelse_(
-        attr(dformula, "lfactor")$correlated,
+        vars$lfactor_defs$correlated,
[…]
