### 2022-12-12 ac7982b
more tests for latent factors and lfo
