### 2022-12-12 8b82a59
lfactor etc defs as attributes to model_vars
