+++ .github/workflows/test-coverage.yaml
@@ -27,7 +27,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          cache-version: 3
+          cache-version: 1
           extra-packages: |
             covr
             BH
