+++ .github/workflows/R-CMD-check_cranrel.yaml
@@ -1,6 +1,9 @@
 # Workflow derived from https://github.com/r-lib/actions/tree/master/examples
 # This is as CMD-check workflow of dynamite
 # but uses CRAN version of rstan and StanHeaders
+# Currently skip this as we need dev version of data.table anyway, and Stan on
+# CRAN is outdated.
+
 on:
   push:
     branches: [main, master]
@@ -47,16 +50,16 @@ jobs:
 
       - uses: r-lib/actions/setup-r@v2
         with:
-          extra-repositories: |
-            https://mc-stan.org/r-packages/
-            https://Rdatatable.gitlab.io/data.table
+          #extra-repositories: |
+          #  https://mc-stan.org/r-packages/
+          #  https://Rdatatable.gitlab.io/data.table
           r-version: ${{ matrix.config.r }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          cache-version: 4
+          cache-version: 1
           extra-packages: |
             rcmdcheck
             BH
