+++ .github/workflows/R-CMD-check.yaml
@@ -57,7 +57,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          cache-version: 3
+          cache-version: 1
           extra-packages: |
             rcmdcheck
             BH
