### 2022-12-14 946e70c
ropensci refs, add reviewers to DESC
