+++ README.md
@@ -8,9 +8,9 @@
 [![Project Status: Active – The project has reached a stable, usable
 state and is being actively
 developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![R-CMD-check](https://github.com/santikka/dynamite/workflows/R-CMD-check/badge.svg)](https://github.com/santikka/dynamite/actions)
+[![R-CMD-check](https://github.com/ropensci/dynamite/workflows/R-CMD-check/badge.svg)](https://github.com/ropensci/dynamite/actions)
 [![Codecov test
-coverage](https://codecov.io/gh/santikka/dynamite/branch/main/graph/badge.svg)](https://app.codecov.io/gh/santikka/dynamite?branch=main)
+coverage](https://codecov.io/gh/ropensci/dynamite/branch/main/graph/badge.svg)](https://app.codecov.io/gh/ropensci/dynamite?branch=main)
 [![Status at rOpenSci Software Peer
 Review](https://badges.ropensci.org/554_status.svg)](https://github.com/ropensci/software-review/issues/554)
 <!-- badges: end -->
@@ -21,22 +21,22 @@ series) data comprising of multiple measurements per multiple
 individuals measured in time. The main features distinguishing the
 package and the underlying methodology from many other approaches are:
 
-- Support for both time-invariant and time-varying effects modeled via
-  B-splines.
-- Joint modeling of multiple measurements per individual (multiple
-  channels) based directly on the assumed data generating process.
-- Support for non-Gaussian observations: Currently Gaussian,
-  Categorical, Poisson, Bernoulli, Binomial, Negative Binomial, Gamma,
-  Exponential, and Beta distributions are available and these can be
-  mixed arbitrarily in multichannel models.
-- Allows evaluating realistic long-term counterfactual predictions which
-  take into account the dynamic structure of the model by posterior
-  predictive distribution simulation.
-- Transparent quantification of parameter and predictive uncertainty due
-  to a fully Bayesian approach.
-- User-friendly and efficient R interface with state-of-the-art
-  estimation via Stan. Both `rstan` and `cmdstanr` backends are
-  supported.
+-   Support for both time-invariant and time-varying effects modeled via
+    B-splines.
+-   Joint modeling of multiple measurements per individual (multiple
+    channels) based directly on the assumed data generating process.
+-   Support for non-Gaussian observations: Currently Gaussian,
+    Categorical, Poisson, Bernoulli, Binomial, Negative Binomial, Gamma,
+    Exponential, and Beta distributions are available and these can be
+    mixed arbitrarily in multichannel models.
+-   Allows evaluating realistic long-term counterfactual predictions
+    which take into account the dynamic structure of the model by
+    posterior predictive distribution simulation.
+-   Transparent quantification of parameter and predictive uncertainty
+    due to a fully Bayesian approach.
+-   User-friendly and efficient R interface with state-of-the-art
+    estimation via Stan. Both `rstan` and `cmdstanr` backends are
+    supported.
 
 The `dynamite` package is developed with the support of Academy of
 Finland grant 331817 ([PREDLIFE](https://sites.utu.fi/predlife/en/)).
@@ -56,11 +56,13 @@ data.table::update_dev_pkg()
 ```
 
 After these steps, you can install the development version of `dynamite`
-from [GitHub](https://github.com/) with:
+from [GitHub](https://github.com/) by running one of the following
+lines:
 
 ``` r
 # install.packages("devtools")
-devtools::install_github("santikka/dynamite")
[…]
