+++ man/gaussian_example.Rd
@@ -16,7 +16,7 @@ A data frame with 3000 rows and 5 variables:
 }
 \source{
 The data was generated according to a script in
-\url{https://github.com/santikka/dynamite/blob/main/data-raw/
+\url{https://github.com/ropensci/dynamite/blob/main/data-raw/
 gaussian_example.R}
 }
 \usage{
