### 2022-12-14 dbd0256
revert data.table ver, run-extended
