+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dynamite
 Title: Bayesian Modeling and Causal Inference for Multivariate
     Longitudinal Data
-Version: 0.0.3
+Version: 1.0.0
 Authors@R: c(
     person("Santtu", "Tikka", email = "santtuth@gmail.com", 
            role = c("aut", "cre"), comment = c(ORCID = "0000-0003-4039-4342")),
@@ -14,7 +14,8 @@ Description: Easy-to-use and efficient interface for
   Bayesian inference of complex panel (time series) data. The package supports 
   joint modeling of multiple measurements per individual, time-varying and
   time-invariant effects, and a wide range of discrete and 
-  continuous distributions. Estimation of the models is carried out via 'Stan'.
+  continuous distributions. Estimation of these dynamic multivariate panel 
+  models is carried out via 'Stan'.
 License: GPL (>= 3)
 URL: https://docs.ropensci.org/dynamite, https://github.com/ropensci/dynamite
 BugReports: https://github.com/ropensci/dynamite/issues
@@ -54,5 +55,6 @@ Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd",
 RoxygenNote: 7.2.3
 LazyData: true
 LazyDataCompression: xz
-Additional_repositories: https://mc-stan.org/r-packages/,
-    https://Rdatatable.gitlab.io/data.table
+Remotes: 
+    url::https://mc-stan.org/r-packages/,
+    url::https://Rdatatable.gitlab.io/data.table
