+++ DESCRIPTION
@@ -55,6 +55,6 @@ Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd",
 RoxygenNote: 7.2.3
 LazyData: true
 LazyDataCompression: xz
+Additional_repositories: https://mc-stan.org/r-packages/
 Remotes: 
-    url::https://mc-stan.org/r-packages/,
     url::https://Rdatatable.gitlab.io/data.table
