+++ .github/workflows/test-coverage.yaml
@@ -36,6 +36,7 @@ jobs:
             Rcpp
             rstan
             StanHeaders
+            r-lib/pak
 
 # To be added when we have tests for cmdstanr backend
 # - name: Build CmdStan
