### 2022-12-14 51e2667
update links to ropensci, refactor additional model component parsing
