### 2022-12-15 99874be
fix citation, remove dep on dev/pak
