+++ .github/workflows/R-CMD-check.yaml
@@ -66,7 +66,6 @@ jobs:
             Rcpp
             rstan
             StanHeaders
-            r-lib/pak
 
 # To be added when we have tests for cmdstanr backend
 # - name: Build CmdStan
