### 2022-12-16 c74e4e2
minor tweak to docs
