+++ R/summary.R
@@ -10,7 +10,8 @@
 #' @return `summary` returns a `data.frame`.
 #' @srrstats {BS6.4, RE4.18} Implements `summary` method.
 #' @examples
-#' summary(gaussian_example_fit, types = "beta")
+#' summary(gaussian_example_fit, types = "beta",
+#'   probs = c(0.05, 0.1, 0.9, 0.95))
 #'
 summary.dynamitefit <- function(object, ...) {
   as.data.frame.dynamitefit(object, ...)
