### 2022-12-19 883abdd
extract correct channels in as_data_table for corr pars
