+++ R/as_data_table.R
@@ -265,7 +265,7 @@ as_data_table_xi <- function(x, draws, ...) {
 #' @describeIn as_data_table_default Data Table for a "corr_nu" Parameter
 #' @noRd
 as_data_table_corr_nu <- function(x, draws, ...) {
-  resp <- get_responses(x$dformulas$stoch)
+  resp <- attr(x$dformulas$stoch, "random")
   pairs <- apply(utils::combn(resp, 2L), 2L, paste, collapse = "_")
   data.table::data.table(
     parameter = paste0("corr_nu_", pairs),
@@ -516,7 +516,7 @@ as_data_table_omega_psi <- function(x, draws, n_draws, category, ...) {
 #' @describeIn as_data_table_default Data Table for a "corr_psi" Parameter
 #' @noRd
 as_data_table_corr_psi <- function(x, draws, ...) {
-  resp <- get_responses(x$dformulas$stoch)
+  resp <- attr(x$dformulas$stoch, "lfactor")
   pairs <- apply(utils::combn(resp, 2L), 2L, paste, collapse = "_")
   data.table::data.table(
     parameter = paste0("corr_psi_", pairs),
