### 2022-12-21 5b8b709
check availability of data.table env argument
