+++ tests/testthat/test-errors.R
@@ -125,13 +125,15 @@ test_that("simultaneity fails", {
     )
   )
   # should fail for deterministic as well
-  expect_error(
-    obs(y ~ x, family = "gaussian") + aux(integer(x) ~ y),
-    paste0(
-      "Simultaneous regression is not supported:\n",
-      "x Response variable `x` appears in the formula of `y`\\."
+  if (datatable_supports_env()) {
+    expect_error(
+      obs(y ~ x, family = "gaussian") + aux(integer(x) ~ y),
+      paste0(
+        "Simultaneous regression is not supported:\n",
+        "x Response variable `x` appears in the formula of `y`\\."
+      )
     )
-  )
+  }
 })
 
 test_that("adding nondynamiteformula to dynamiteformula fails", {
@@ -157,9 +159,9 @@ test_that("plus method fails for nondynamiteformula", {
 test_that("categorical random intercept fails", {
   expect_error(
     dynamite(obs(y ~ x, family = "categorical") + random(),
-      data = data.frame(y = factor(1:4), x = runif(4), id = 1, time = 1:4),
-      "id", "time",
-      debug = list(no_compile = TRUE)
+             data = data.frame(y = factor(1:4), x = runif(4), id = 1, time = 1:4),
+             "id", "time",
+             debug = list(no_compile = TRUE)
     ),
     paste0(
       "No valid responses for random intercept component:\n",
@@ -232,6 +234,7 @@ test_that("lb_tau definition throws error if not of correct length", {
 })
 
 test_that("pure deterministic formula to dynamite fails", {
+  skip_if(!datatable_supports_env())
   expect_error(
     dynamite(
       dformula = aux(numeric(d) ~ lag(d, 1)),
@@ -246,9 +249,9 @@ test_that("pure deterministic formula to dynamite fails", {
 test_that("categorical latent factor fails", {
   expect_error(
     dynamite(obs(y ~ x, family = "categorical") + lfactor(),
-      data = data.frame(y = factor(1:4), x = runif(4), id = 1, time = 1:4),
-      "id", "time",
-      debug = list(no_compile = TRUE)
+             data = data.frame(y = factor(1:4), x = runif(4), id = 1, time = 1:4),
+             "id", "time",
+             debug = list(no_compile = TRUE)
     ),
     paste0(
       "No valid responses for latent factor component:\n",
@@ -260,17 +263,17 @@ test_that("categorical latent factor fails", {
 test_that("Latent factor errors with invalid responses", {
   expect_error(
     dynamite(obs(y ~ x, family = "gaussian") + lfactor(responses = 1),
-      data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
-      "id", "time",
-      debug = list(no_compile = TRUE)
+             data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
+             "id", "time",
+             debug = list(no_compile = TRUE)
     ),
     "Argument `responses` must be a <character> vector\\."
   )
   expect_error(
     dynamite(obs(y ~ x, family = "gaussian") + lfactor(responses = "x"),
-      data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
-      "id", "time",
-      debug = list(no_compile = TRUE)
+             data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
+             "id", "time",
+             debug = list(no_compile = TRUE)
     ),
     paste0(
       "Argument `responses` of `lfactor\\(\\)` contains variable `x`:\n",
@@ -280,43 +283,43 @@ test_that("Latent factor errors with invalid responses", {
 })
[…]
