+++ tests/testthat/test-edgecases.R
@@ -187,8 +187,8 @@ test_that("shrinkage is handled correctly", {
   expect_equal(
     unname(unlist(lapply(
       dynamite(obs_all_alpha, test_data,
-        "group", "time",
-        debug = debug
+               "group", "time",
+               debug = debug
       )$stan$model_vars, "[[", "shrinkage"
     ))),
     rep(TRUE, 6)
@@ -214,8 +214,8 @@ test_that("noncentered splines are handled correctly", {
   expect_equal(
     unname(unlist(lapply(
       dynamite(obs_all_alpha, test_data,
-        "group", "time",
-        debug = debug
+               "group", "time",
+               debug = debug
       )$stan$model_vars, "[[", "noncentered"
     ))),
     rep(TRUE, 6)
@@ -236,8 +236,8 @@ test_that("lower bounds for tau are handled correctly", {
   expect_equal(
     lapply(
       dynamite(obs_all_alpha, test_data,
-        "group", "time",
-        debug = debug
+               "group", "time",
+               debug = debug
       )$stan$model_vars, "[[", "lb"
     ),
     list(y1 = 0, y2 = 2, y3 = 1, y4 = 0.4, y9 = 0.01)
@@ -257,7 +257,7 @@ test_that("latent factors are handled correctly", {
       obs(y3 ~ -1 + x3 + varying(~x1) + trials(trials), family = "binomial") +
       obs(y4 ~ x1 + varying(~ -1 + x2), family = "bernoulli") +
       splines(df = 5) + lfactor(c("y2", "y3"),
-        nonzero_lambda = c(TRUE, FALSE, FALSE)),
+                                nonzero_lambda = c(TRUE, FALSE, FALSE)),
     NA
   )
   expect_equal(
@@ -277,7 +277,7 @@ test_that("lags are parsed", {
   )
   expect_error(
     obs_b <- obs(y1 ~ -1 + x1 + varying(~ lag(y2, 1)),
-      family = "categorical"
+                 family = "categorical"
     ) +
       obs(y2 ~ -1 + x2 + varying(~ lag(y1, 1)), family = "gaussian") +
       splines(),
@@ -337,9 +337,9 @@ test_that("higher order lags() and lag() give equal results", {
     obs(y2 ~ x2, family = "gaussian") +
     lags(k = 1:2, type = "fixed")
   f2 <- obs(y1 ~ x1 + lag(y1, 1) + lag(y2, 1) +
-    lag(y1, 2) + lag(y2, 2), family = "categorical") +
+              lag(y1, 2) + lag(y2, 2), family = "categorical") +
     obs(y2 ~ x2 + lag(y1, 1) + lag(y2, 1) +
-      lag(y1, 2) + lag(y2, 2), family = "gaussian")
+          lag(y1, 2) + lag(y2, 2), family = "gaussian")
   expect_identical(
     get_priors(f1, test_data, "group", "time"),
     get_priors(f2, test_data, "group", "time")
@@ -442,9 +442,10 @@ test_that("no groups data preparation works", {
 # Deterministic edgecases -------------------------------------------------
 
 test_that("deterministic channels are parsed", {
+  skip_if(!datatable_supports_env())
   expect_error(
     obs_det <- obs(y5 ~ x1 + lag(d, 1) + lag(y5, 1) + lag(x1, 1),
-      family = "negbin"
+                   family = "negbin"
     ) +
       aux(numeric(d) ~ lag(d, 1) + lag(f, 2) + x2 | init(0)) +
       aux(numeric(f) ~ lag(y5, 1) + x2 * 3 + 1 | init(c(0, 1))),
[…]

[…]
