### 2022-12-21 1fac834
deterministic channels without env
