+++ tests/testthat/test-errors.R
@@ -125,15 +125,13 @@ test_that("simultaneity fails", {
     )
   )
   # should fail for deterministic as well
-  if (datatable_supports_env()) {
-    expect_error(
-      obs(y ~ x, family = "gaussian") + aux(integer(x) ~ y),
-      paste0(
-        "Simultaneous regression is not supported:\n",
-        "x Response variable `x` appears in the formula of `y`\\."
-      )
+  expect_error(
+    obs(y ~ x, family = "gaussian") + aux(integer(x) ~ y),
+    paste0(
+      "Simultaneous regression is not supported:\n",
+      "x Response variable `x` appears in the formula of `y`\\."
     )
-  }
+  )
 })
 
 test_that("adding nondynamiteformula to dynamiteformula fails", {
@@ -234,7 +232,6 @@ test_that("lb_tau definition throws error if not of correct length", {
 })
 
 test_that("pure deterministic formula to dynamite fails", {
-  skip_if_not(datatable_supports_env())
   expect_error(
     dynamite(
       dformula = aux(numeric(d) ~ lag(d, 1)),
@@ -344,7 +341,6 @@ test_that("binomial channel without a trials term fails", {
 })
 
 test_that("deterministic fixed fails", {
-  skip_if_not(datatable_supports_env())
   expect_error(
     aux(numeric(y) ~ fixed(~x)),
     paste0(
@@ -357,7 +353,6 @@ test_that("deterministic fixed fails", {
 })
 
 test_that("deterministic varying fails", {
-  skip_if_not(datatable_supports_env())
   expect_error(
     aux(numeric(y) ~ varying(~x)),
     paste0(
@@ -483,7 +478,6 @@ test_that("missing predictor fails", {
 })
 
 test_that("invalid deterministic channel definition fails", {
-  skip_if_not(datatable_supports_env())
   expect_error(
     dynamite(
       dformula = obs(y ~ x, family = "gaussian") +
