### 2022-12-21 93b73ef
modified multichannel example data, removed grunfeld example due to size limits
