+++ vignettes/dynamite.Rmd
@@ -368,64 +368,9 @@ of the data). Arguments `chains` and `cores` are passed to `rstan::sampling`
 which then uses to parallel Markov chains in the the Markov chain Monte Carlo 
 (MCMC) sampling of the model parameters.
 
-# Examples
+# Example of Multichannel Model
 
-## Grunfeld Data
-
-As an example, we consider the famous Grunfeld dataset, available in the 
-package `plm`:
-
-```{r}
-library(dynamite)
-library(dplyr)
-data(Grunfeld, package = "plm")
-d <- Grunfeld |> 
-  mutate(invest = inv, firm = factor(firm)) |> 
-  select(-inv)
-head(d)
-```
-
-The data is in a long format (as required by the `dynamite`), containing the 
-firm id (`firm`, 10 firms), the `year` (20 years), value of the firm (`value`), 
-the firms’ gross investment (`invest`) and stock of plant and equipment 
-(`capital`). 
-
-Let's first fit the basic model without time-invariant effects of value and 
-capital on the investment, but with a firm-specific random intercept 
-(using the `random_intercept = TRUE` inside of `obs`):
-
-```{r, eval = FALSE}
-fit_grunfeld <- dynamite(obs(invest ~ value + capital,
-  family = "gaussian") + random(),
-  d, "firm", "year",
-  chains = 1, cores = 1, refresh = 500, iter = 2000, 
-  save_warmup = FALSE)
-fit_grunfeld
-```
-```{r, echo = FALSE}
-fit_grunfeld <- readRDS("fit_grunfeld.rds")
-fit_grunfeld
-```
-
-
-Instead of random intercept terms for each firm, an fixed effect approach would 
-have been possible by using the formula 
-`obs(invest ~ firm + value + capital, family = "gaussian")`.
-
-As we used weakly informative default priors, the coefficient estimates are 
-close to the ones obtained with the `plm` package which is based on  
-ordinary least squares:
-
-```{r}
-fit_plm <- plm::plm(invest ~ value + capital, data = d, 
-                    index = c("firm", "year"), 
-                    effect = "individual", model = "within")
-fit_plm
-```
-
-## Multichannel Model
-
-As a second example, consider the following simulated multichannel data:
+As an example, consider the following simulated multichannel data:
 
 ```{r}
 library(dynamite)
@@ -448,8 +393,11 @@ f <- obs(g ~ lag(g) + lag(logp), family = "gaussian") +
 As a directed acyclic graph (DAG), the model for the first three time points is 
 ![A DAG of the multichannel model.](dag.png){width=80%}
 
-We again fit the model using the `dynamite` function (note the small number of 
-iterations and additional thinning in order to satisfy the CRAN requirements):
+(The deterministic channel `logp` is omitted for clarity).
+
+We fit the model using the `dynamite` function. In order to satisfy the package 
+size requirements of CRAN, we have to use small number of 
+iterations and additional thinning:
 
 ```{r, eval = FALSE}
 multichannel_example_fit <- dynamite(
@@ -458,7 +406,20 @@ multichannel_example_fit <- dynamite(
   thin = 5, save_warmup = FALSE)
 ```
 
-Note that this produces a warning message related to the 
+You can update the estimated model which is 
+included in the package as 
+```{r, eval = FALSE}
+multichannel_example_fit <- update(multichannel_example_fit,
+    iter = 2000,
+    warmup = 1000,
+    chains = 4,
+    cores = 4,
+    refresh = 0,
+    save_warmup = FALSE
+)
+```
+
+Note that the `dynamite` call above produces a warning message related to the 
[…]
