+++ DESCRIPTION
@@ -42,7 +42,6 @@ Suggests:
     covr,
     dplyr,
     knitr,
-    plm,
     rmarkdown,
     testthat (>= 3.0.0),
     tidyr
