### 2022-12-21 377dfb5
output type change, tries to fix #54
