+++ vignettes/dynamite.Rmd
@@ -1,9 +1,7 @@
 ---
 title: "dynamite"
 link-citations: yes
-output:
-  html_document: default
-  pdf_document: default
+output: rmarkdown::html_vignette
 bibliography: dynamite.bib
 vignette: >
   %\VignetteIndexEntry{dynamite} 
