### 2022-12-21 d5ff4b7
fix some urls, fix lfo, remove data.table additional repo, add doi to DESCRIPTION, run-extended
