+++ vignettes/dynamite.bib
@@ -1,6 +1,6 @@
 @article{Bollen2010,
  ISSN = {00377732, 15347605},
- URL = {http://www.jstor.org/stable/40927552},
+ URL = {https://www.jstor.org/stable/40927552},
  author = {Kenneth A. Bollen and Jennie E. Brand},
  journal = {Social Forces},
  number = {1},
@@ -30,7 +30,7 @@
 }
 @article{Holtz-Eakin1988,
  ISSN = {00129682, 14680262},
- URL = {http://www.jstor.org/stable/1913103},
+ URL = {https://www.jstor.org/stable/1913103},
  author = {Douglas Holtz-Eakin and Whitney Newey and Harvey S. Rosen},
  journal = {Econometrica},
  number = {6},
