### 2022-12-21 cc19e81
reduce the number of samples to keep the package smaller than 5mb
