### 2022-12-21 392d89b
use mathjax in docs
