+++ NEWS.md
@@ -6,6 +6,7 @@
     decreased in order to make the example in the vignette more interesting.
   * The latent factor model was also modified by removing the `random()` 
     component in order to reduce the size of the model fit object.
+  * Fixed the name extraction of the supplied data.
 
 # dynamite 1.0.0
   * Increased the version number to 1.0.0 to reflect the fact that the package 
