### 2022-12-23 cbf5a3e
use donttest in examples, fix na warning in plot_deltas, update reference in DESC
