+++ R/dynamiteformula.R
@@ -273,7 +273,7 @@ aux <- function(formula) {
 #' @examples
 #' x <- obs(y ~ x, family = "gaussian") +
 #'   obs(z ~ w, family = "exponential") +
-#'   aux(d ~ log(y) | init(c(0, 1))) +
+#'   aux(numeric(d) ~ log(y) | init(c(0, 1))) +
 #'   lags(k = 2) +
 #'   splines(df = 5) +
 #'   random(responses = c("y", "z"), correlated = TRUE)
