+++ man/plot.lfo.Rd
@@ -19,8 +19,9 @@ Plots Pareto k values per each time point (with one point per group),
 together with the horizontal line representing the used threshold.
 }
 \examples{
-\dontrun{
-plot(lfo(gaussian_example_fit, L = 20))
+\donttest{
+# this gives warnings due to the small number of iterations
+plot(suppressWarnings(lfo(gaussian_example_fit, L = 20)))
 }
 
 }

[…]
