### 2022-12-27 e1f59d7
local gamma_ to gamma__ to avoid name clashes, fix #55
