+++ R/stanblocks_families.R
@@ -623,13 +623,13 @@ transformed_parameters_lines_default <- function(y, idt, noncentered,
     state_omega_alpha_1 <- paste_rows(
       "// Define the first alpha using mean a_{y}",
       "{{",
-      "vector[{K}] gamma_{y};",
-      onlyif(has_fixed, "gamma_{y}[{{{cs(L_fixed)}}}] = beta_{y};"),
+      "vector[{K}] gamma__{y};",
+      onlyif(has_fixed, "gamma__{y}[{{{cs(L_fixed)}}}] = beta_{y};"),
       onlyif(
         has_varying,
-        "gamma_{y}[{{{cs(L_varying)}}}] = delta_{y}[1];"
+        "gamma__{y}[{{{cs(L_varying)}}}] = delta_{y}[1];"
       ),
-      "omega_alpha_1_{y} = a_{y} - X_m[{{{cs(J)}}}] * gamma_{y};",
+      "omega_alpha_1_{y} = a_{y} - X_m[{{{cs(J)}}}] * gamma__{y};",
       "}}",
       .indent = idt(c(1, 1, 2, 2, 2, 2, 1)),
       .parse = FALSE
@@ -643,13 +643,13 @@ transformed_parameters_lines_default <- function(y, idt, noncentered,
     state_fixed_intercept <- paste_rows(
       "// Define the first alpha using mean a_{y}",
       "{{",
-      "vector[{K}] gamma_{y};",
-      onlyif(has_fixed, "gamma_{y}[{{{cs(L_fixed)}}}] = beta_{y};"),
+      "vector[{K}] gamma__{y};",
+      onlyif(has_fixed, "gamma__{y}[{{{cs(L_fixed)}}}] = beta_{y};"),
       onlyif(
         has_varying,
-        "gamma_{y}[{{{cs(L_varying)}}}] = delta_{y}[1];"
+        "gamma__{y}[{{{cs(L_varying)}}}] = delta_{y}[1];"
       ),
-      "alpha_{y} = a_{y} - X_m[{{{cs(J)}}}] * gamma_{y};",
+      "alpha_{y} = a_{y} - X_m[{{{cs(J)}}}] * gamma__{y};",
       "}}",
       .indent = idt(c(1, 1, 2, 2, 2, 2, 1)),
       .parse = FALSE
@@ -805,16 +805,16 @@ transformed_parameters_lines_categorical <- function(y, idt, noncentered,
     state_omega_alpha_1 <- paste_rows(
       "// Define the first alpha using mean a_{y}",
       "for (s in 1:{S - 1}) {{",
-      "vector[{K}] gamma_{y};",
+      "vector[{K}] gamma__{y};",
       onlyif(
         has_fixed,
-        "gamma_{y}[{{{cs(L_fixed)}}}] = beta_{y}[, s];"
+        "gamma__{y}[{{{cs(L_fixed)}}}] = beta_{y}[, s];"
       ),
       onlyif(
         has_varying,
-        "gamma_{y}[{{{cs(L_varying)}}}] = delta_{y}[1, , s];"
+        "gamma__{y}[{{{cs(L_varying)}}}] = delta_{y}[1, , s];"
       ),
-      "omega_alpha_1_{y}[s] = a_{y}[s] - X_m[{{{cs(J)}}}] * gamma_{y};",
+      "omega_alpha_1_{y}[s] = a_{y}[s] - X_m[{{{cs(J)}}}] * gamma__{y};",
       "}}",
       .indent = idt(c(1, 1, 2, 2, 2, 2, 1)),
       .parse = FALSE
[…]
