### 2022-12-27 74d2aa9
update NEWS, increment ver, add badges
