+++ README.md
@@ -13,6 +13,10 @@ developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.re
 coverage](https://codecov.io/gh/ropensci/dynamite/branch/main/graph/badge.svg)](https://app.codecov.io/gh/ropensci/dynamite?branch=main)
 [![Status at rOpenSci Software Peer
 Review](https://badges.ropensci.org/554_status.svg)](https://github.com/ropensci/software-review/issues/554)
+[![dynamite status
+badge](https://ropensci.r-universe.dev/badges/dynamite)](https://ropensci.r-universe.dev)
+[![dynamite
+cran-badge](http://www.r-pkg.org/badges/version/dynamite)](https://cran.r-project.org/package=dynamite)
 <!-- badges: end -->
 
 The `dynamite` [R](https://www.r-project.org/) package provides
@@ -21,22 +25,22 @@ series) data comprising of multiple measurements per multiple
 individuals measured in time. The main features distinguishing the
 package and the underlying methodology from many other approaches are:
 
-- Support for both time-invariant and time-varying effects modeled via
-  B-splines.
-- Joint modeling of multiple measurements per individual (multiple
-  channels) based directly on the assumed data generating process.
-- Support for non-Gaussian observations: Currently Gaussian,
-  Categorical, Poisson, Bernoulli, Binomial, Negative Binomial, Gamma,
-  Exponential, and Beta distributions are available and these can be
-  mixed arbitrarily in multichannel models.
-- Allows evaluating realistic long-term counterfactual predictions which
-  take into account the dynamic structure of the model by posterior
-  predictive distribution simulation.
-- Transparent quantification of parameter and predictive uncertainty due
-  to a fully Bayesian approach.
-- User-friendly and efficient R interface with state-of-the-art
-  estimation via Stan. Both `rstan` and `cmdstanr` backends are
-  supported.
+-   Support for both time-invariant and time-varying effects modeled via
+    B-splines.
+-   Joint modeling of multiple measurements per individual (multiple
+    channels) based directly on the assumed data generating process.
+-   Support for non-Gaussian observations: Currently Gaussian,
+    Categorical, Poisson, Bernoulli, Binomial, Negative Binomial, Gamma,
+    Exponential, and Beta distributions are available and these can be
+    mixed arbitrarily in multichannel models.
+-   Allows evaluating realistic long-term counterfactual predictions
+    which take into account the dynamic structure of the model by
+    posterior predictive distribution simulation.
+-   Transparent quantification of parameter and predictive uncertainty
+    due to a fully Bayesian approach.
+-   User-friendly and efficient R interface with state-of-the-art
+    estimation via Stan. Both `rstan` and `cmdstanr` backends are
+    supported.
 
 The `dynamite` package is developed with the support of Academy of
 Finland grant 331817 ([PREDLIFE](https://sites.utu.fi/predlife/en/)).
@@ -82,6 +86,7 @@ Posterior estimates of time-varying effects
 
 ``` r
 plot_deltas(gaussian_example_fit, scales = "free")
+#> Warning: Removed 1 row containing missing values (`geom_line()`).
 ```
 
 <img src="man/figures/README-unnamed-chunk-6-1.png" width="80%" />
@@ -108,6 +113,7 @@ first time point):
 
 ``` r
 library(ggplot2)
+#> Warning: package 'ggplot2' was built under R version 4.2.2
 pred <- predict(gaussian_example_fit, n_draws = 50)
 pred |> dplyr::filter(id < 5) |> 
   ggplot(aes(time, y_new, group = .draw)) +
@@ -124,29 +130,29 @@ For more examples, see the package vignette.
 
 ## Related packages
 
[…]
