### 2023-01-11 4223cb1
add the dynamite logo
