### 2023-01-13 e6420f8
fixed a typo in fitted docs
