### 2023-01-16 309588c
some additions to docs and refs to LOO/LFO
