### 2023-01-18 4fe1b62
model with random effects compiles
