+++ R/stanblocks_families.R
@@ -34,6 +34,8 @@ vectorizable_prior <- function(x) {
 #'   predictors?
 #' @param has_varying \[`logical(1)`]\cr Does the channel have time-varying
 #'   predictors?
+#' @param has_random \[`logical(1)`]\cr  Does the channel have random
+#' (group-specific) effects?
 #' @param has_fixed_intercept \[`logical(1)`]\cr Does the channel have a
 #'   time-invariant intercept?
 #' @param has_varying_intercept \[`logical(1)`]\cr Does the channel have a
@@ -55,6 +57,8 @@ vectorizable_prior <- function(x) {
 #' @param K_fixed \[`integer(1)`]\cr Number of time-invariant predictors of
 #'   the channel.
 #' @param K_varying \[`integer(1)`]\cr Number of time-varying predictors of
+#'   the channel.
+#' @param K_random \[`integer(1)`]\cr Number of random effect predictors of
 #'   the channel.
 #' @param L_fixed \[`integer(1)`]\cr Indices of the time-invariant predictors
 #'   of the channel in the joint parameter vector `gamma`.
@@ -376,7 +380,6 @@ transformed_data_lines_beta <- function(...) {
 parameters_lines_default <- function(y, idt, noncentered, lb, has_fixed,
                                      has_varying, has_fixed_intercept,
                                      has_varying_intercept,
-                                     has_random_intercept,
                                      K_fixed, K_varying,
                                      has_lfactor, noncentered_psi,
                                      noncentered_lambda,
@@ -388,10 +391,6 @@ parameters_lines_default <- function(y, idt, noncentered, lb, has_fixed,
       predictors contain possibly nonzero latent factor."))
   }
   paste_rows(
-    onlyif(
-      has_random_intercept,
-      "real<lower=0> sigma_nu_{y}; // SD of random intercepts"
-    ),
     onlyif(has_fixed, "vector[{K_fixed}] beta_{y}; // Fixed coefficients"),
     onlyif(
       has_varying,
@@ -1002,7 +1001,8 @@ transformed_parameters_lines_beta <- function(...) {
 # Model block -------------------------------------------------------------
 
 model_lines_default <- function(y, idt, obs, noncentered, shrinkage,
-                                has_varying, has_fixed, has_fixed_intercept,
+                                has_varying, has_fixed, has_random,
+                                has_fixed_intercept,
                                 has_varying_intercept, has_random_intercept,
                                 has_lfactor, noncentered_psi,
                                 noncentered_lambda, nonzero_lambda,
@@ -1019,7 +1019,7 @@ model_lines_default <- function(y, idt, obs, noncentered, shrinkage,
                                 sigma_lambda_prior_distr = "",
                                 psi_prior_distr = "",
                                 tau_psi_prior_distr = "",
-                                K_fixed, K_varying, ...) {
+                                K_fixed, K_varying, K_random, J_random, ...) {
 
   allow_intercept <- !(has_lfactor && nonzero_lambda)
 
@@ -1154,8 +1154,8 @@ model_lines_default <- function(y, idt, obs, noncentered, shrinkage,
     has_random_intercept,
[…]
