### 2023-01-18 66379fe
tweak parsing of random_spec
