### 2023-01-18 de2f27e
lower bound to sigma_nu
