### 2023-01-19 d89d136
print of random effect specification
