### 2023-01-19 54661d1
fix default sigma_nu prior defs
