+++ tests/testthat/test-recovery.R
@@ -88,6 +88,46 @@ test_that("parameters for the gamma glm are recovered as with glm", {
   )
 })
 
+test_that("parameters for poisson mixed model are recovered", {
+  skip_if_not(run_extended_tests)
+  set.seed(1)
+  n <- 40
+  k <- 10
+  x <- rnorm(n * k)
+  u1 <- rep(rnorm(k, sd = 0.2), each = n)
+  u2 <- rep(rnorm(k, sd = 0.1), each = n)
+  y <- rpois(n * k, exp(2 - x + u1 + u2 * x))
+  d <- data.frame(year = 1:n, person = rep(1:k, each = n), y = y, x = x)
+
+  # use default priors of brms (except not totally flat for beta)
+  p <- data.frame(
+    parameter = c("sigma_nu_y_alpha", "sigma_nu_y_x", "alpha_y", "beta_y_x",
+      "L_nu"),
+    response = "y",
+    prior = c("student_t(3, 0, 2.5)", "student_t(3, 0, 2.5)",
+      "student_t(3, 1.9, 2.5)", "normal(0, 1000)", "lkj_corr_cholesky(1)"),
+    type = c("sigma_nu", "sigma_nu", "alpha", "beta", "L"),
+    category = ""
+  )
+  fit_dynamite <- dynamite(obs(y ~ x + random(~1 + x), family = "poisson") +
+      random_spec(noncentered = FALSE, correlated = TRUE),
+    data = d, time = "year", group = "person", priors = p,
+    chains = 1, iter = 2000, refresh = 0
+  )
+  # "ground truth" obtained from one long dynamite run and brms,
+  # note that brms can give few divergences as does dynamite if
+  # noncentered = TRUE
+  expect_equal(coef(fit_dynamite)$mean, c(2.014, -0.9932),
+    tolerance = 0.05
+  )
+  expect_equal(coef(fit_dynamite, type="nu")$mean,
+    c(0.1635, 0.4153, -0.0924, -0.1344, -0.0773, -0.1237, -0.2027,
+      -0.1231, 0.2726, -0.1071, -0.03, 4e-04, 0.0997, -0.1146, -0.0313,
+      0.0146, 0.0407, -0.0169, -0.1407, 0.1635),
+    tolerance = 0.05
+  )
+})
+
 test_that("parameters for an AR(1) model are recovered as with arima", {
   skip_if_not(run_extended_tests)
   set.seed(1)
