+++ data-raw/gaussian_example_fit.R
@@ -8,10 +8,10 @@ set.seed(1)
 gaussian_example_fit <- dynamite(
   dformula =
     obs(
-      y ~ -1 + z + varying(~ x + lag(y)),
+      y ~ -1 + z + varying(~ x + lag(y)) + random(~1),
       family = "gaussian"
     ) +
-      random() +
+      random_spec() +
       splines(df = 20),
   data = gaussian_example,
   group = "id",
@@ -23,7 +23,8 @@ gaussian_example_fit <- dynamite(
   cores = 2,
   refresh = 0,
   save_warmup = FALSE,
-  pars = c("omega_alpha_1_y", "omega_raw_alpha_y", "nu_raw", "nu", "L"),
+  pars = c("omega_alpha_1_y", "omega_raw_alpha_y", "nu_raw", "nu", "L",
+    "sigma_nu", "a_y"),
   include = FALSE
 )
 
