+++ R/utilities.R
@@ -77,26 +77,49 @@ drop_unused <- function(dformula, data, group_var, time_var) {
 #' @param x \[`character()`]\cr A vector of terms to add.
 #' @param type \[`character(1)`]\cr Either `"fixed"` or `"varying"`
 #'   indicating the type of terms to add.
-#' @param varying_idx \[`integer()`] Indices of left-hand side terms that have
-#'   time-varying coefficients
-#' @param varying_icpt \[`logical(1)`] Does the formula have a varying
+#' @param varying_idx \[`integer()`]\cr Indices of left-hand side terms that
+#'   have time-varying coefficients.
+#' @param random_idx \[`integer()`]\cr Indices of the left-hand side term that
+#'   have random coefficient terms.
+#' @param varying_icpt \[`logical(1)`]\cr Does the formula have a varying
 #'   intercept?
-#' @param fixed_icpt \[`logical(1)`] Does the formula have a fixed intercept?
+#' @param fixed_icpt \[`logical(1)`]\cr
+#'   Does the formula have a fixed intercept?
+#' @param random_icpt \[`logical(1)`]\cr
+#'   Does the formula have a random intercept
 #' @noRd
-increment_formula <- function(formula, x, type = c("fixed", "varying", "random"),
-                              varying_idx, varying_icpt, fixed_icpt) {
+increment_formula <- function(formula, x,
+                              type = c("fixed", "varying", "random"),
+                              varying_idx, random_idx,
+                              varying_icpt, fixed_icpt, random_icpt) {
   v_icpt <- ifelse_(varying_icpt, "1", "-1")
+  r_icpt <- ifelse_(random_icpt, "1", "-1")
   n_varying <- length(varying_idx)
+  n_random <- length(random_idx)
   x_plus <- paste0(x, collapse = " + ")
   ft <- terms(formula)
   tr <- attr(ft, "term.labels")
   v <- paste0(tr[varying_idx], collapse = " + ")
+  r <- paste0(tr[random_idx], collapse = " + ")
   formula_str <- ""
   if (n_varying > 0L) {
     if (n_varying < length(tr)) {
       formula <- stats::drop.terms(
         ft,
         dropx = varying_idx,
+        keep.response = TRUE
+      )
+      ft <- terms(formula)
+      tr <- attr(ft, "term.labels")
+    } else {
+      tr <- character(0L)
+    }
+  }
+  if (n_random > 0L) {
+    if (n_random < length(tr)) {
+      formula <- stats::drop.terms(
+        ft,
+        dropx = random_idx,
         keep.response = TRUE
       )
       ft <- terms(formula)
@@ -119,17 +142,32 @@ increment_formula <- function(formula, x, type = c("fixed", "varying", "random")
       ifelse_(fixed_icpt, "1", "-1")
     )
   }
-  if (identical(type, "varying")) {
-    v <- paste0(" + ", v, ifelse_(nzchar(v), " + ", ""), x_plus)
-    out_str <- glue::glue("{formula_str} + varying(~{v_icpt}{v})")
-  } else {
-    v <- ifelse_(nzchar(v), paste0(" + ", v), v)
-    out_str <- ifelse_(
-      n_varying > 0L || varying_icpt,
-      glue::glue("{formula_str} + {x_plus} + varying(~{v_icpt}{v})"),
-      glue::glue("{formula_str} + {x_plus}")
-    )
-  }
+  v_out <- ifelse_(
+    identical(type, "varying"),
[…]
