### 2023-01-19 fea45e8
fix prior vectorization for sigma_nu, refit gaussian_example
