+++ R/stanblocks_families.R
@@ -65,14 +65,17 @@ vectorizable_prior <- function(x) {
 #' @param L_varying \[`integer(1)`]\cr Indices of the time-varying predictors
 #'   of the channel in the joint parameter vector `gamma`.
 #' @param S \[`integer(1)`]\cr Number of categories for a categorical channel.
-#' @param write_alpha \[`logical(1)`]\cr Should the `alpha` parameters of the
-#'   model be written to the model code?
-#' @param write_beta \[`logical(1)`]\cr Should the `beta` parameters of
-#'   time-invariant predictors be written to the model code?
-#' @param write_delta \[`logical(1)`]\cr Should the `delta` parameters of
-#'   time-varying predictors be written to the model code?
-#' @param write_tau \[`logical(1)`]\cr Should the `tau` parameters be written
-#'   to the model code?
+#' @param write_alpha \[`logical(1)`]\cr If `TRUE`, use vectorized prior for
+#'   `alpha` parameters of the model by hard-coding the parameters of the prior
+#'   distribution to the model code?
+#' @param write_beta \[`logical(1)`]\cr If `TRUE`, use vectorized prior for
+#'   time-invariant predictors `beta`.
+#' @param write_delta \[`logical(1)`]\cr If `TRUE`, use vectorized prior for
+#'   time-varying predictors `delta`.
+#' @param write_tau \[`logical(1)`]\cr If `TRUE`, use vectorized prior for
+#'   `tau` parameters.
+#' @param write_sigma_nu \[`logical(1)`]\cr If `TRUE`, use vectorized prior for
+#'   `sigma_nu` parameters.
 #' @param sigma_prior_distr \[`character(1)`]\cr `sigma` parameter prior
 #'   specification.
 #' @param sigma_nu_prior_distr \[`character(1)`]\cr `sigma_nu` parameter prior
@@ -94,6 +97,10 @@ vectorizable_prior <- function(x) {
 #' @param tau_prior_distr \[`character(1)`]\cr `tau` parameter prior
 #'   specification.
 #' @param tau_prior_npars \[`integer(1)`]\cr Number of parameters for the
+#'   prior of `tau`.
+#' @param sigma_nu_prior_distr \[`character(1)`]\cr `sigma_nu` parameter prior
+#'   specification.
+#' @param sigma_nu_prior_npars \[`integer(1)`]\cr Number of parameters for the
 #'   prior of `tau`.
 #' @param phi_prior_distr \[`character(1)`]\cr `phi` parameter prior
 #'   specification.
@@ -200,13 +207,16 @@ data_lines_beta <- function(y, idt, has_missing, ...) {
 # Transformed data block --------------------------------------------------
 
 transformed_data_lines_default <- function(y, idt, write_beta, write_delta,
-                                           write_tau, K_fixed, K_varying,
+                                           write_tau, write_sigma_nu, K_fixed,
+                                           K_varying, K_random,
                                            beta_prior_npars = 1L,
                                            beta_prior_pars = "",
                                            delta_prior_npars = 1L,
                                            delta_prior_pars = "",
                                            tau_prior_npars = 1L,
-                                           tau_prior_pars = "") {
+                                           tau_prior_pars = "",
+                                           sigma_nu_prior_npars = 1L,
[…]
