### 2023-01-19 7c04706
predict with general random effects (new groups not yet supported)
