+++ R/predict_helpers.R
@@ -348,34 +348,34 @@ clear_nonfixed <- function(newdata, newdata_null, resp_stoch, eval_type,
   }
 }
 
-#' Generate Random Intercepts for New Group Levels
+#' Generate Random Effects for New Group Levels
 #'
 #' @inheritParams predict
-#' @param nu Posterior draws of the random intercept parameters.
+#' @param nu Posterior draws of the random effects.
 #' @param sigma_nu Posterior draws of the standard deviations of the random
-#'   intercept parameters.
+#'   effects.
 #' @param corr_matrix_nu Posterior draws of the correlation matrix of
-#'   intercepts (within-group).
+#'   random effects.
 #' @param n_group \[`integer(1)`]\cr Number of groups.
 #' @param orig_ids \[`character()`]\cr Group levels of the original data.
 #' @param new_ids \[`character()`]\cr Group levels of `newdata` in
 #'   [dynamite::predict.dynamitefit()].
 #' @param new_levels \[`character(1)`]\cr
-#'   Defines if and how to sample the random intercepts for observations whose
-#'   group level was not present in the original data. The options are
+#'   Defines if and how to sample the random effects for groups not present in
+#'   the original data. The options are
 #'     * `"none"` (the default) which will signal an error if new levels
 #'       are encountered.
 #'     * `"bootstrap"` which will randomly draw from the posterior samples of
-#'       the random intercepts across all original levels.
+#'       the random effects across all original levels.
 #'     * `"gaussian"` which will randomly draw from a gaussian
-#'       distribution using the posterior samples of the random intercept
-#'       standard deviation (and correlation matrix if applicable).
+#'       distribution using the posterior samples of the random effect
+#'       standard deviations (and correlation matrix if applicable).
 #'     * `"original"` which will randomly match each new level to one of
-#'       the original levels. The posterior samples of the random intercept of
+#'       the original levels. The posterior samples of the random effects of
 #'       the matched levels will then be used for the new levels.
 #' @return An n_draws x n_groups x n_intercepts array of random intercepts.
 #' @noRd
-generate_random_intercept <- function(nu, sigma_nu, corr_matrix_nu, n_draws,
+generate_random_effect <- function(nu, sigma_nu, corr_matrix_nu, n_draws,
                                       n_group, orig_ids, new_ids, new_levels) {
   is_orig <- which(orig_ids %in% new_ids)
   is_new <- !new_ids %in% orig_ids
@@ -456,10 +456,9 @@ prepare_eval_envs <- function(object, simulated, observed,
   n_resp <- length(resp_stoch)
   eval_envs <- vector(mode = "list", length = n_resp)
   idx_draws <- seq_len(n_draws)
-  nu_channels <- attr(object$dformulas$stoch, "random")$responses
-  M <- length(nu_channels)
+  nu_channels <- which_random(object$dformulas$all)
   n_group <- n_unique(observed[[group_var]])
-  if (M > 0L) {
+  if (length(nu_channels) > 0L) {
     orig_ids <- unique(object$data[[group_var]])
     new_ids <- unique(observed[[group_var]])
     n_all_draws <- ndraws(object)
@@ -468,18 +467,19 @@ prepare_eval_envs <- function(object, simulated, observed,
       do.call("cbind", samples[sigma_nus])[idx_draws, , drop = FALSE]
     )
     nus <- glue::glue("nu_{nu_channels}")
[…]
