### 2023-01-20 cbb4c71
fix extraction of correlation parameters
