### 2023-01-20 ade66d6
check random effects only from stochastic channels
