### 2023-01-20 ce3b44c
refit example models, fix predict tests
