### 2023-01-20 cd81250
remove bayesplot, fix dynamiteformula example
