+++ man/dynamiteformula.Rd
@@ -119,7 +119,7 @@ or time-varying predictors via \code{\link[=lags]{lags()}} instead of writing th
 for each channel.
 
 It is also possible to define group-specific (random) effects term
-using the special function \code{\link[=random]{random()}} similarly as \code{\link[=varying]{varying()}}. For example,
+using the special syntax \code{random()} similarly as \code{varying()}. For example,
 \code{random(~1)} leads to a model where in addition to the common intercept,
 each individual/group has their own intercept with zero-mean normal prior and
 unknown standard deviation analogously with the typical mixed models. An
@@ -156,7 +156,7 @@ obs(g ~ lag(g) + lag(logp), family = "gaussian") +
 
 obs(y ~ x, family = "gaussian") + obs(z ~ w, family = "exponential")
 
-x <- obs(y ~ x + random(~1 + d), family = "gaussian") +
+x <- obs(y ~ x + random(~1 + lag(d)), family = "gaussian") +
   obs(z ~ varying(~w), family = "exponential") +
   aux(numeric(d) ~ log(y) | init(c(0, 1))) +
   lags(k = 2) +
