+++ DESCRIPTION
@@ -23,7 +23,6 @@ BugReports: https://github.com/ropensci/dynamite/issues/
 Depends:
     R (>= 3.5.0)
 Imports: 
-    bayesplot,
     checkmate,
     cli,
     data.table,
@@ -32,6 +31,7 @@ Imports:
     loo,
     MASS,
     methods,
+    patchwork,
     posterior,
     rlang,
     rstan,
