+++ NAMESPACE
@@ -53,6 +53,7 @@ export(splines)
 importFrom(data.table,":=")
 importFrom(data.table,as.data.table)
 importFrom(loo,loo)
+importFrom(patchwork,wrap_plots)
 importFrom(posterior,as_draws)
 importFrom(posterior,as_draws_df)
 importFrom(posterior,ndraws)
