+++ tests/testthat/test-errors.R
@@ -158,7 +158,7 @@ test_that("categorical with random effects fails", {
   expect_error(
     dynamite(obs(y ~ x + random(~1), family = "categorical") + random_spec(),
       data = data.frame(y = factor(1:4), x = runif(4), id = 1, time = 1:4),
-      "id", "time",
+      "time", "id",
       debug = list(no_compile = TRUE)
     ),
     "Random effects are not \\(yet\\) supported for categorical responses."
@@ -200,7 +200,7 @@ test_that("noncentered definition throws error if not of correct length", {
     NA
   )
   expect_error(
-    dynamite(obs_all_alpha, test_data, "group", "time"),
+    dynamite(obs_all_alpha, test_data, "time", "group"),
     paste(
       "Length of the `noncentered` argument of `splines\\(\\)` function",
       "is not equal to 1 or 6, the number of the channels\\."
@@ -220,7 +220,7 @@ test_that("lb_tau definition throws error if not of correct length", {
     NA
   )
   expect_error(
-    dynamite(obs_all_alpha, test_data, "group", "time"),
+    dynamite(obs_all_alpha, test_data, "time", "group"),
     paste(
       "Length of the `lb_tau` argument of `splines\\(\\)` function is not",
       "equal to 1 or 6, the number of the channels\\."
@@ -242,10 +242,12 @@ test_that("pure deterministic formula to dynamite fails", {
 
 test_that("categorical latent factor fails", {
   expect_error(
-    dynamite(obs(y ~ x, family = "categorical") + lfactor(),
-             data = data.frame(y = factor(1:4), x = runif(4), id = 1, time = 1:4),
-             "id", "time",
-             debug = list(no_compile = TRUE)
+    dynamite(
+      obs(y ~ x, family = "categorical") + lfactor(),
+      data = data.frame(y = factor(1:4), x = runif(4), id = 1, time = 1:4),
+      time = "time",
+      group = "id",
+      debug = list(no_compile = TRUE)
     ),
     paste0(
       "No valid responses for latent factor component:\n",
@@ -256,18 +258,22 @@ test_that("categorical latent factor fails", {
 
 test_that("Latent factor errors with invalid responses", {
   expect_error(
-    dynamite(obs(y ~ x, family = "gaussian") + lfactor(responses = 1),
-             data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
-             "id", "time",
-             debug = list(no_compile = TRUE)
+    dynamite(
+      obs(y ~ x, family = "gaussian") + lfactor(responses = 1),
+      data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
+      time = "time",
+      group = "id",
+      debug = list(no_compile = TRUE)
     ),
     "Argument `responses` must be a <character> vector\\."
   )
   expect_error(
-    dynamite(obs(y ~ x, family = "gaussian") + lfactor(responses = "x"),
-             data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
-             "id", "time",
-             debug = list(no_compile = TRUE)
+    dynamite(
+      obs(y ~ x, family = "gaussian") + lfactor(responses = "x"),
+      data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
+      time = "time",
+      group = "id",
+      debug = list(no_compile = TRUE)
     ),
     paste0(
       "Argument `responses` of `lfactor\\(\\)` contains variable `x`:\n",
@@ -278,10 +284,12 @@ test_that("Latent factor errors with invalid responses", {
 test_that("Latent factor errors with nonlogical value for argument
   noncentered", {
     expect_error(
[…]
