+++ tests/testthat/test-edgecases.R
@@ -59,31 +59,31 @@ test_that("single channel models are valid", {
     obs_beta <- obs(y9 ~ x3, family = "beta"), NA
   )
   expect_error(
-    dynamite(obs_categorical, test_data, "group", "time", debug = debug), NA
+    dynamite(obs_categorical, test_data, "time", "group", debug = debug), NA
   )
   expect_error(
-    dynamite(obs_gaussian, test_data, "group", "time", debug = debug), NA
+    dynamite(obs_gaussian, test_data, "time", "group", debug = debug), NA
   )
   expect_error(
-    dynamite(obs_binomial, test_data, "group", "time", debug = debug), NA
+    dynamite(obs_binomial, test_data, "time", "group", debug = debug), NA
   )
   expect_error(
-    dynamite(obs_bernoulli, test_data, "group", "time", debug = debug), NA
+    dynamite(obs_bernoulli, test_data, "time", "group", debug = debug), NA
   )
   expect_error(
-    dynamite(obs_negbinom, test_data, "group", "time", debug = debug), NA
+    dynamite(obs_negbinom, test_data, "time", "group", debug = debug), NA
   )
   expect_error(
-    dynamite(obs_poisson, test_data, "group", "time", debug = debug), NA
+    dynamite(obs_poisson, test_data, "time", "group", debug = debug), NA
   )
   expect_error(
-    dynamite(obs_exponential, test_data, "group", "time", debug = debug), NA
+    dynamite(obs_exponential, test_data, "time", "group", debug = debug), NA
   )
   expect_error(
-    dynamite(obs_gamma, test_data, "group", "time", debug = debug), NA
+    dynamite(obs_gamma, test_data, "time", "group", debug = debug), NA
   )
   expect_error(
-    dynamite(obs_beta, test_data, "group", "time", debug = debug), NA
+    dynamite(obs_beta, test_data, "time", "group", debug = debug), NA
   )
 })
 
@@ -101,7 +101,7 @@ test_that("multichannel models are valid", {
     NA
   )
   expect_error(
-    dynamite(obs_all, test_data, "group", "time", debug = debug), NA
+    dynamite(obs_all, test_data, "time", "group", debug = debug), NA
   )
 })
 
@@ -110,7 +110,7 @@ test_that("binomial with trials only works", {
     obs_binomial <- obs(y3 ~ trials(trials), family = "binomial"), NA
   )
   expect_error(
-    dynamite(obs_binomial, test_data, "group", "time", debug = debug), NA
+    dynamite(obs_binomial, test_data, "time", "group", debug = debug), NA
   )
 })
 
@@ -124,7 +124,7 @@ test_that("intercepts are handled correctly", {
     NA
   )
   expect_error(
-    dynamite(obs_all_alpha, test_data, "group", "time", debug = debug), NA
+    dynamite(obs_all_alpha, test_data, "time", "group", debug = debug), NA
   )
 })
 
@@ -138,7 +138,7 @@ test_that("random effects are handled correctly", {
   )
   expect_equal(
     c("sigma_nu_y2_alpha", "sigma_nu_y3_alpha", "L_nu"),
-    get_priors(obs_all_alpha, test_data, "group", "time")$parameter[c(1, 6, 24)]
+    get_priors(obs_all_alpha, test_data, "time", "group")$parameter[c(1, 6, 24)]
   )
 
   expect_error(
@@ -181,7 +181,7 @@ test_that("random effects are handled correctly", {
     NA
   )
   expect_false(
-    "L_nu" %in% get_priors(obs_all_alpha, test_data, "group", "time")$parameter
[…]
