### 2023-01-24 2014cbe
correctly parse formulas with random factors
