+++ R/model_matrix.R
@@ -4,16 +4,36 @@
 #' @srrstats {RE1.3, RE1.3a} `full_model.matrix` preserves relevant attributes.
 #' @noRd
 full_model.matrix <- function(dformula, data, verbose) {
-  formulas <- get_formulas(dformula)
-  model_matrices <- vector(mode = "list", length = length(formulas))
-  for (i in seq_along(formulas)) {
-    mm <-
-      stats::model.matrix.lm(formulas[[i]], data = data, na.action = na.pass)
+  #formulas <- get_formulas(dformula)
+  model_matrices <- vector(mode = "list", length = length(dformula))
+  model_matrices_type <- vector(mode = "list", length = length(dformula))
+  types <- c("fixed", "varying", "random")
+  for (i in seq_along(dformula)) {
+    mm <- stats::model.matrix.lm(
+      dformula[[i]]$formula,
+      data = data,
+      na.action = na.pass
+    )
     if (verbose) {
       test_collinearity(dformula[[i]]$resp, mm, data)
     }
+    #formulas <- stats::setNames(vector(mode = "list", length = 3L), types)
+    model_matrices_type[[i]] <- list()
+    for (type in c("fixed", "varying", "random")) {
+      type_formula <- get_type_formula(dformula[[i]], type)
+      if (!is.null(type_formula)) {
+        model_matrices_type[[i]][[type]] <- remove_intercept(
+          stats::model.matrix.lm(
+            type_formula,
+            data = data,
+            na.action = na.pass
+          )
+        )
+      }
+    }
+    model_matrices[[i]] <- do.call(cbind, model_matrices_type[[i]])
     # Intercept is not part of X
-    model_matrices[[i]] <- remove_intercept(mm)
+    #model_matrices[[i]] <- remove_intercept(mm)
   }
   model_matrix <- do.call(cbind, model_matrices)
   u_names <- unique(colnames(model_matrix))
@@ -27,17 +47,34 @@ full_model.matrix <- function(dformula, data, verbose) {
   attr(model_matrix, "fixed") <- empty_list
   attr(model_matrix, "varying") <- empty_list
   attr(model_matrix, "random") <- empty_list
+  model_df <- as.data.frame(model_matrix)
   for (i in seq_along(model_matrices)) {
-    cols <- colnames(model_matrices[[i]])
-    assign <- match(cols, u_names)
-    assign_i <- attr(model_matrices[[i]], "assign")
-    attr(model_matrix, "assign")[[i]] <- sort(assign)
-    fixed <- assign[assign_i %in% dformula[[i]]$fixed]
-    attr(model_matrix, "fixed")[[i]] <- setNames(fixed, u_names[fixed])
-    varying <- assign[assign_i %in% dformula[[i]]$varying]
-    attr(model_matrix, "varying")[[i]] <- setNames(varying, u_names[varying])
-    random <- assign[assign_i %in% dformula[[i]]$random]
-    attr(model_matrix, "random")[[i]] <- setNames(random, u_names[random])
+    for (type in types) {
+      if (!is.null(model_matrices_type[[i]][[type]])) {
+        model_df_type <- as.data.frame(model_matrices_type[[i]][[type]])
+        cols <- which(model_df %in% model_df_type)
+        attr(model_matrix, type)[[i]] <- setNames(cols, u_names[cols])
+      } else {
+        attr(model_matrix, type)[[i]] <- integer(0L)
+      }
+    }
+    attr(model_matrix, "assign")[[i]] <- sort(
+      c(
+        attr(model_matrix, "fixed")[[i]],
+        attr(model_matrix, "varying")[[i]],
+        attr(model_matrix, "random")[[i]]
+      )
[…]
