### 2023-01-24 df192fc
narrower output of print.dynamitefit
