### 2023-01-24 648a30e
J_fixed etc and priors as data to stan
