+++ R/stanblocks_families.R
@@ -109,164 +109,169 @@ NULL
 
 # Data block --------------------------------------------------------------
 
-data_lines_default <- function(y, idt, has_missing, ...) {
+data_lines_default <- function(y, idt, has_missing,
+                               write_beta, write_delta, write_tau,
+                               write_sigma_nu,
+                               beta_prior_npars = 1L, delta_prior_npars = 1L,
+                               tau_prior_npars = 1L, sigma_nu_prior_npars = 1L,
+                               ...) {
   paste_rows(
     onlyif(has_missing, "// Missing data indicators"),
     onlyif(has_missing, "int<lower=0> obs_{y}[N, T];"),
     onlyif(has_missing, "int<lower=0> n_obs_{y}[T];"),
+    "int<lower=0> K_fixed_{y};",
+    "int<lower=0> K_varying_{y};",
+    "int<lower=0> K_random_{y};",
+    "int<lower=0> K_{y};",
+    "int<lower=0> J_fixed_{y};",
+    "int<lower=0> J_varying_{y};",
+    "int<lower=0> J_random_{y};",
+    "int<lower=0> J_{y};",
+    "int<lower=0> L_fixed_{y};",
+    "int<lower=0> L_varying_{y};",
+    "int<lower=0> L_random_{y};",
+    onlyif(write_beta,
+      "matrix[K_fixed_{y}, {beta_prior_npars}] beta_prior_pars_{y}"),
+    onlyif(write_delta,
+      "matrix[K_varying_{y}, {delta_prior_npars}] delta_prior_pars_{y}"),
+    onlyif(write_tau,
+      "matrix[K_varying_{y}, {tau_prior_npars}] tau_prior_pars_{y};"),
+    onlyif(write_sigma_nu,
+      "matrix[K_random_{y}, {sigma_nu_prior_npars}] sigma_nu_prior_pars_{y};"),
     "// Data",
     .indent = idt(1)
   )
 }
 
-data_lines_categorical <- function(y, idt, has_missing, ...) {
-  dtext_def <- data_lines_default(y, idt, has_missing)
-  dtext <- paste_rows(
+
+data_lines_categorical <- function(y, idt,
+                                   has_missing, write_alpha, write_beta,
+                                   write_delta, write_tau,
+                                   alpha_prior_npars = 1L,
+                                   beta_prior_npars = 1L,
+                                   delta_prior_npars = 1L, tau_prior_npars = 1L,
+                                   ...) {
+  dtext_def <- paste_rows(
+    "int<lower=0> S_{y}; // number of categories",
+    onlyif(has_missing, "// Missing data indicators"),
+    onlyif(has_missing, "int<lower=0> obs_{y}[N, T];"),
+    onlyif(has_missing, "int<lower=0> n_obs_{y}[T];"),
+    "int<lower=0> K_fixed_{y};",
+    "int<lower=0> K_varying_{y};",
+    "int<lower=0> K_random_{y};",
+    "int<lower=0> K_{y};",
+    "int<lower=0> J_fixed_{y};",
+    "int<lower=0> J_varying_{y};",
+    "int<lower=0> J_random_{y};",
+    "int<lower=0> J_{y};",
+    "int<lower=0> L_fixed_{y};",
+    "int<lower=0> L_varying_{y};",
+    "int<lower=0> L_random_{y};",
+    onlyif(write_alpha,
+      "matrix[S_{y} - 1, {alpha_prior_npars}] alpha_prior_pars_{y};"),
+    onlyif(write_beta,
[…]
