+++ R/stanblocks_families.R
@@ -131,9 +131,9 @@ data_lines_default <- function(y, idt, has_missing,
     "int<lower=0> L_varying_{y};",
     "int<lower=0> L_random_{y};",
     onlyif(write_beta,
-      "matrix[K_fixed_{y}, {beta_prior_npars}] beta_prior_pars_{y}"),
+      "matrix[K_fixed_{y}, {beta_prior_npars}] beta_prior_pars_{y};"),
     onlyif(write_delta,
-      "matrix[K_varying_{y}, {delta_prior_npars}] delta_prior_pars_{y}"),
+      "matrix[K_varying_{y}, {delta_prior_npars}] delta_prior_pars_{y};"),
     onlyif(write_tau,
       "matrix[K_varying_{y}, {tau_prior_npars}] tau_prior_pars_{y};"),
     onlyif(write_sigma_nu,
