### 2023-01-25 be49273
some comments to generated stan code
