+++ R/stanblocks_families.R
@@ -119,17 +119,20 @@ data_lines_default <- function(y, idt, has_missing,
     onlyif(has_missing, "// Missing data indicators"),
     onlyif(has_missing, "int<lower=0> obs_{y}[N, T];"),
     onlyif(has_missing, "int<lower=0> n_obs_{y}[T];"),
+    "// number of fixed, varying and random coefficients, and related indices",
     "int<lower=0> K_fixed_{y};",
     "int<lower=0> K_varying_{y};",
-    "int<lower=0> K_random_{y};",
-    "int<lower=0> K_{y};",
+    "int<lower=0> K_random_{y}; // includes the potential random intercept",
+    "int<lower=0> K_{y}; // K_fixed + K_varying",
     "int<lower=0> J_fixed_{y};",
     "int<lower=0> J_varying_{y};",
     "int<lower=0> J_random_{y};",
-    "int<lower=0> J_{y};",
+    "int<lower=0> J_{y}; // fixed and varying",
     "int<lower=0> L_fixed_{y};",
     "int<lower=0> L_varying_{y};",
     "int<lower=0> L_random_{y};",
+    onlyif(write_beta || write_delta || write_tau || write_sigma_nu,
+      "// Parameters of vectorized priors"),
     onlyif(write_beta,
       "matrix[K_fixed_{y}, {beta_prior_npars}] beta_prior_pars_{y};"),
     onlyif(write_delta,
@@ -138,7 +141,7 @@ data_lines_default <- function(y, idt, has_missing,
       "matrix[K_varying_{y}, {tau_prior_npars}] tau_prior_pars_{y};"),
     onlyif(write_sigma_nu,
       "matrix[K_random_{y}, {sigma_nu_prior_npars}] sigma_nu_prior_pars_{y};"),
-    "// Data",
+    "// Responses",
     .indent = idt(1)
   )
 }
@@ -156,17 +159,20 @@ data_lines_categorical <- function(y, idt,
     onlyif(has_missing, "// Missing data indicators"),
     onlyif(has_missing, "int<lower=0> obs_{y}[N, T];"),
     onlyif(has_missing, "int<lower=0> n_obs_{y}[T];"),
+    "// number of fixed, varying and random coefficients, and related indices",
     "int<lower=0> K_fixed_{y};",
     "int<lower=0> K_varying_{y};",
-    "int<lower=0> K_random_{y};",
-    "int<lower=0> K_{y};",
+    "int<lower=0> K_random_{y}; // includes the potential random intercept",
+    "int<lower=0> K_{y}; // K_fixed + K_varying",
     "int<lower=0> J_fixed_{y};",
     "int<lower=0> J_varying_{y};",
     "int<lower=0> J_random_{y};",
-    "int<lower=0> J_{y};",
+    "int<lower=0> J_{y}; // fixed and varying",
     "int<lower=0> L_fixed_{y};",
     "int<lower=0> L_varying_{y};",
     "int<lower=0> L_random_{y};",
+    onlyif(write_alpha || write_beta || write_delta || write_tau,
+      "// Parameters of vectorized priors"),
     onlyif(write_alpha,
       "matrix[S_{y} - 1, {alpha_prior_npars}] alpha_prior_pars_{y};"),
     onlyif(write_beta,
@@ -175,7 +181,7 @@ data_lines_categorical <- function(y, idt,
       "matrix[K_varying_{y} * (S_{y} - 1), {delta_prior_npars}] delta_prior_pars_{y};"),
     onlyif(write_delta,
       "matrix[K_varying_{y}, {tau_prior_npars}] tau_prior_pars_{y};"),
-    "// Data",
+    "// Responses",
     "int<lower=0> y_{y}[T, N];",
     .indent = idt(1)
   )
