### 2023-01-25 d2aff55
pass indices etc correctly to stan
