### 2023-01-25 df4dec7
as.array to L_fixed etc
