### 2023-01-25 0da1874
evaluate model matrix in inner loop
