### 2023-01-25 cc51458
full sequential contemporaneous dependency support including deterministic channels
