### 2023-01-26 aac3207
allow acyclic contemporaneous formulas, update fit objects
