+++ R/stanblocks_families.R
@@ -136,16 +136,26 @@ data_lines_default <- function(y, idt, has_missing,
     "int L_fixed_{y}[K_fixed_{y}];",
     "int L_varying_{y}[K_varying_{y}];",
     "int L_random_{y}[K_random_{y}{icpt}];",
-    onlyif(write_beta || write_delta || write_tau || write_sigma_nu,
-      "// Parameters of vectorized priors"),
-    onlyif(write_beta,
-      "matrix[K_fixed_{y}, {beta_prior_npars}] beta_prior_pars_{y};"),
-    onlyif(write_delta,
-      "matrix[K_varying_{y}, {delta_prior_npars}] delta_prior_pars_{y};"),
-    onlyif(write_tau,
-      "matrix[K_varying_{y}, {tau_prior_npars}] tau_prior_pars_{y};"),
-    onlyif(write_sigma_nu,
-      "matrix[K_random_{y}, {sigma_nu_prior_npars}] sigma_nu_prior_pars_{y};"),
+    onlyif(
+      write_beta || write_delta || write_tau || write_sigma_nu,
+      "// Parameters of vectorized priors"
+    ),
+    onlyif(
+      write_beta,
+      "matrix[K_fixed_{y}, {beta_prior_npars}] beta_prior_pars_{y};"
+    ),
+    onlyif(
+      write_delta,
+      "matrix[K_varying_{y}, {delta_prior_npars}] delta_prior_pars_{y};"
+    ),
+    onlyif(
+      write_tau,
+      "matrix[K_varying_{y}, {tau_prior_npars}] tau_prior_pars_{y};"
+    ),
+    onlyif(
+      write_sigma_nu,
+      "matrix[K_random_{y}, {sigma_nu_prior_npars}] sigma_nu_prior_pars_{y};"
+    ),
     "// Responses",
     .indent = idt(1)
   )
@@ -157,7 +167,7 @@ data_lines_categorical <- function(y, idt,
                                    alpha_prior_npars = 1L,
                                    beta_prior_npars = 1L,
                                    delta_prior_npars = 1L, tau_prior_npars = 1L,
-  has_random_intercept, ...) {
+                                   has_random_intercept, ...) {
   icpt <- ifelse(
     has_random_intercept,
     " - 1",
@@ -180,16 +190,26 @@ data_lines_categorical <- function(y, idt,
     "int L_fixed_{y}[K_fixed_{y}];",
     "int L_varying_{y}[K_varying_{y}];",
     "int L_random_{y}[K_random_{y}{icpt}];",
-    onlyif(write_alpha || write_beta || write_delta || write_tau,
-      "// Parameters of vectorized priors"),
-    onlyif(write_alpha,
-      "matrix[S_{y} - 1, {alpha_prior_npars}] alpha_prior_pars_{y};"),
-    onlyif(write_beta,
-      "matrix[K_fixed_{y} * (S_{y} - 1), {beta_prior_npars}] beta_prior_pars_{y};"),
-    onlyif(write_delta,
-      "matrix[K_varying_{y} * (S_{y} - 1), {delta_prior_npars}] delta_prior_pars_{y};"),
-    onlyif(write_delta,
-      "matrix[K_varying_{y}, {tau_prior_npars}] tau_prior_pars_{y};"),
+    onlyif(
+      write_alpha || write_beta || write_delta || write_tau,
+      "// Parameters of vectorized priors"
+    ),
+    onlyif(
+      write_alpha,
+      "matrix[S_{y} - 1, {alpha_prior_npars}] alpha_prior_pars_{y};"
+    ),
+    onlyif(
+      write_beta,
[…]
