### 2023-01-26 a78cd6e
remove dims from sigma and phi
