### 2023-01-26 2242bd5
tweak printing of mcmc diagnostics
