### 2023-01-27 96a0630
switch from check_hmc_diagnostics to get_ functions
