### 2023-01-27 561dae2
deterministic predict fix (summary)
