### 2023-01-29 b6db376
support std_normal prior
