### 2023-01-30 0d1a5c7
fix splines in formula.dynamitefit
