### 2023-01-30 27a6612
compute M using model.matrix
