### 2023-01-30 d3399b4
Add lfactor and random_spec to formula.dynamitefit
