+++ R/dynamite.R
@@ -441,8 +441,8 @@ formula.dynamitefit <- function(x, ...) {
     function(y) y$name,
     character(1L)
   )
-  lag_defs <- attr(x$dformulas$all, "lags")
-  spline_defs <- attr(x$dformulas$stoch, "splines")
+  lag_def <- attr(x$dformulas$all, "lags")
+  spline_def <- attr(x$dformulas$stoch, "splines")
   obs_str <- onlyif(
     length(ch_stoch) > 0L,
     obs_str <- paste0(
@@ -460,22 +460,55 @@ formula.dynamitefit <- function(x, ...) {
     )
   )
   lags_str <- onlyif(
-    !is.null(lag_defs),
-    glue::glue("lags(k = {lag_defs$k}, type = {lag_defs$type})")
+    !is.null(lag_def),
+    glue::glue("lags(k = {lag_def$k}, type = {lag_def$type})")
   )
   spline_str <- onlyif(
-    spline_defs$has_splines,
+    spline_def$has_splines,
     paste0(
       "splines(",
-      "shrinkage = ", spline_defs$shrinkage, ", ",
+      "shrinkage = ", spline_def$shrinkage, ", ",
       "override = FALSE, ",
-      "df = ", spline_defs$bs_opts$df, ", ",
-      "degree = ", spline_defs$bs_opts$degree, ", ",
-      "lb_tau = ", spline_defs$lb_tau, ", ",
-      "noncentered = ", spline_defs$noncentered, ")"
+      "df = ", spline_def$bs_opts$df, ", ",
+      "degree = ", spline_def$bs_opts$degree, ", ",
+      "lb_tau = ", spline_def$lb, ", ",
+      "noncentered = ", spline_def$noncentered, ")"
     )
   )
-  str2lang(paste(c(obs_str, aux_str, lags_str, spline_str), collapse = " + "))
+  lfactor_def <- attr(x$dformulas$stoch, "lfactor")
+  lfactor_str <- onlyif(
+    lfactor_def$has_lfactor,
+    paste0(
+      "lfactor(",
+      "responses = ", lfactor_def$responses, ", ",
+      "noncentered_psi = ", lfactor_def$noncentered_psi, ", ",
+      "noncentered_lambda = ", lfactor_def$noncentered_lambda, ", ",
+      "nonzero_lambda = ", lfactor_def$nonzero_lambda, ", ",
+      "correlated = ", lfactor_def$correlated, ")"
+    )
+  )
+  random_spec_def <- attr(x$dformulas$stoch, "random_spec")
+  random_spec_str <- onlyif(
+    random_spec_def$has_random_spec,
+    paste0(
+      "random_spec(",
+      "correlated = ", random_spec_def$correlated, ", ",
+      "noncentered = ", random_spec_def$noncentered, ")"
+    )
+  )
+  str2lang(
+    paste(
+      c(
+        obs_str,
+        aux_str,
+        lags_str,
+        spline_str,
+        lfactor_str,
+        random_spec_str
+      ),
+      collapse = " + "
+    )
+  )
 }
 
 #' Is The Argument a `dynamitefit` Object
@@ -731,7 +764,7 @@ parse_components <- function(dformulas, data, time_var) {
   resp <- get_responses(dformulas$stoch)
   families <- unlist(get_families(dformulas$stoch))
   attr(dformulas$stoch, "splines") <- parse_splines(
-    spline_defs = attr(dformulas$stoch, "splines"),
+    spline_def = attr(dformulas$stoch, "splines"),
     resp = resp,
     times = seq.int(fixed + 1L, n_unique(data[[time_var]]))
   )
[…]
