### 2023-01-30 a441c97
remove redundant function
