### 2023-01-30 3aeb728
support std_normal(), fix #56
