+++ R/stanblocks_families.R
@@ -141,19 +141,19 @@ data_lines_default <- function(y, idt, has_missing,
       "// Parameters of vectorized priors"
     ),
     onlyif(
-      write_beta,
-      "matrix[K_fixed_{y}, {beta_prior_npars}] beta_prior_pars_{y};"
+      write_beta && beta_prior_npars > 0L,
+        "matrix[K_fixed_{y}, {beta_prior_npars}] beta_prior_pars_{y};"
     ),
     onlyif(
-      write_delta,
+      write_delta && delta_prior_npars > 0L,
       "matrix[K_varying_{y}, {delta_prior_npars}] delta_prior_pars_{y};"
     ),
     onlyif(
-      write_tau,
+      write_tau && tau_prior_npars > 0L,
       "matrix[K_varying_{y}, {tau_prior_npars}] tau_prior_pars_{y};"
     ),
     onlyif(
-      write_sigma_nu,
+      write_sigma_nu && sigma_nu_prior_npars > 0L,
       "matrix[K_random_{y}, {sigma_nu_prior_npars}] sigma_nu_prior_pars_{y};"
     ),
     "// Responses",
@@ -195,19 +195,19 @@ data_lines_categorical <- function(y, idt,
       "// Parameters of vectorized priors"
     ),
     onlyif(
-      write_alpha,
+      write_alpha && alpha_prior_npars > 0L,
       "matrix[S_{y} - 1, {alpha_prior_npars}] alpha_prior_pars_{y};"
     ),
     onlyif(
-      write_beta,
+      write_beta && beta_prior_npars > 0L,
       "matrix[K_fixed_{y} * (S_{y} - 1), {beta_prior_npars}] beta_prior_pars_{y};"
     ),
     onlyif(
-      write_delta,
+      write_delta && delta_prior_npars > 0L,
       "matrix[K_varying_{y} * (S_{y} - 1), {delta_prior_npars}] delta_prior_pars_{y};"
     ),
     onlyif(
-      write_delta,
+      write_tau && tau_prior_npars > 0L,
       "matrix[K_varying_{y}, {tau_prior_npars}] tau_prior_pars_{y};"
     ),
     "// Responses",
@@ -1020,9 +1020,12 @@ model_lines_default <- function(y, idt, obs, noncentered, shrinkage,
   mtext_alpha <- "a_{y} ~ {alpha_prior_distr};"
 
   if (vectorizable_prior(sigma_nu_prior_distr)) {
-    dpars_sigma_nu <- paste0(
-      "sigma_nu_prior_pars_", y, "[, ", seq_len(sigma_nu_prior_npars), "]",
-      collapse = ", "
+    dpars_sigma_nu <- ifelse_(sigma_nu_prior_npars > 0L,
+      paste0(
+        "sigma_nu_prior_pars_", y, "[, ", seq_len(sigma_nu_prior_npars), "]",
+        collapse = ", "
+      ),
+      ""
     )
     mtext_sigma_nu <- "sigma_nu_{y} ~ {sigma_nu_prior_distr}({dpars_sigma_nu});"
   } else {
@@ -1053,9 +1056,12 @@ model_lines_default <- function(y, idt, obs, noncentered, shrinkage,
   if (noncentered) {
     mtext_omega <- "omega_raw_alpha_{y} ~ std_normal();"
     if (vectorizable_prior(delta_prior_distr)) {
-      dpars_varying <- paste0(
-        "delta_prior_pars_", y, "[, ", seq_len(delta_prior_npars), "]",
-        collapse = ", "
+      dpars_varying <- ifelse_(delta_prior_npars > 0L,
+        paste0(
+          "delta_prior_pars_", y, "[, ", seq_len(delta_prior_npars), "]",
+          collapse = ", "
[…]
