### 2023-01-30 737ee16
remove note on nonlag response for aux
