+++ man/dynamiteformula.Rd
@@ -78,8 +78,7 @@ In addition to declaring response variables via \code{obs}, we can also use
 the function \code{aux} to define auxiliary channels which are deterministic
 functions of other variables. The values of auxiliary variables are computed
 dynamically during prediction, making the use of lagged values and other
-transformations possible. Note that the auxiliary channel can also depend
-on other variables without lags. The function \code{aux} also does not use the
+transformations possible. The function \code{aux} also does not use the
 \code{family} argument, which is automatically set to \code{deterministic} and is a
 special channel type of \code{obs}. Note that lagged values of deterministic
 \code{aux} channels do not imply fixed time points. Instead they must be given
