### 2023-01-30 f3fea08
extended test for predictions
