### 2023-01-30 cc64529
get_code blocks argument
