+++ tests/testthat/test-recovery.R
@@ -331,14 +331,14 @@ test_that("prior parameters are recovered with zero observations", {
 })
 
 test_that("predict recovers correct estimates", {
-
+  skip_if_not(run_extended_tests)
   set.seed(1)
   N <- 20
   T_ <- 30
   y <- matrix(0, N, T_)
   nu <- rnorm(N)
   y[, 1] <- rbinom(N, size = 1, prob = 0.5)
-  for(t in 2:T) y[, t] <- rbinom(N, 1, plogis(nu + y[, t-1]))
+  for(t in 2:T_) y[, t] <- rbinom(N, 1, plogis(nu + y[, t-1]))
 
   ## check these if tests fail ##
   # model <- rstan::stan_model("testmodel.stan")
