### 2023-01-31 3c4ebad
evaluate_deterministic before parse_components
