### 2023-01-31 d667ab8
change defaults and column ordering of summary and as.data.frame
