+++ R/summary.R
@@ -1,12 +1,12 @@
 #' Summary for a Dynamite Model Fit
 #'
 #' The `summary` method provides statistics of the posterior samples of the
-#' model; this is an alias of [dynamite::as.data.frame.dynamitefit()].
+#' model; this is an alias of [dynamite::as.data.frame.dynamitefit()] with
+#' `summary = TRUE`.
 #'
 #' @export
 #' @rdname dynamite
 #' @param object \[`dynamitefit`]\cr The model fit object.
-#' @param ... Further arguments to [dynamite::as.data.frame.dynamitefit()].
 #' @return `summary` returns a `data.frame`.
 #' @srrstats {BS6.4, RE4.18} Implements `summary` method.
 #' @examples
@@ -16,5 +16,5 @@
 #' )
 #'
 summary.dynamitefit <- function(object, ...) {
-  as.data.frame.dynamitefit(object, ...)
+  as.data.frame.dynamitefit(object, summary = TRUE, ...)
 }
