### 2023-01-31 93ed916
remove extra empty line from mcmc_diagnostics
