### 2023-02-01 b795549
prevent multiple fixed/varying/random terms in formula
